refactor(runner): split monolith into domain files and fix shutdown handling
Split the 3,663-line runner.go into 7 files by concern:
- runner.go: core types, interface, orchestration
- lifecycle.go: container lifecycle management
- strategy_container.go: container-recreate rollback strategy
- strategy_checkpoint.go: checkpoint-restore rollback strategy
- rpc.go: RPC and Engine API communication
- logging.go: log streaming, writers, hooks
- resources.go: resource limits, CPU pinning, CPU frequency

Fix Podman CTRL+C handling:
- Use context.Background() for Podman connection so API calls survive
  parent context cancellation (bindings embed ctx in connection object)
- Derive cancellable context for WaitForContainerExit so it unblocks
  on signal instead of hanging until container exits on its own
- Force-remove current container immediately on cancellation in
  container-recreate strategy instead of waiting for deferred cleanup
- Support second CTRL+C to force-exit when graceful shutdown is slow

Fix container-recreate strategy gaps:
- Add pre_run_steps execution before ZFS snapshot and per-test for
  non-ZFS path (was missing, checkpoint-restore had it)
- Honour wait_after_rpc_ready config in both ZFS and non-ZFS paths
- Log "RPC endpoint ready" with client version after per-test container
  startup (matching lifecycle.go's initial container behavior)
- Log latest block info after RPC ready for per-test containers
