+++ pkg/runner/runner.go
@@ -4,37 +4,21 @@ import (
 	"context"
 	"crypto/rand"
 	"encoding/hex"
-	"encoding/json"
-	"errors"
 	"fmt"
-	"io"
-	mrand "math/rand/v2"
-	"net/http"
 	"os"
-	"os/exec"
 	"path/filepath"
-	"strconv"
-	"strings"
 	"sync"
 	"time"
 
 	dockerclient "github.com/docker/docker/client"
-	"github.com/docker/go-units"
 	"github.com/ethpandaops/benchmarkoor/pkg/blocklog"
 	"github.com/ethpandaops/benchmarkoor/pkg/client"
 	"github.com/ethpandaops/benchmarkoor/pkg/config"
 	"github.com/ethpandaops/benchmarkoor/pkg/cpufreq"
-	"github.com/ethpandaops/benchmarkoor/pkg/datadir"
 	"github.com/ethpandaops/benchmarkoor/pkg/docker"
 	"github.com/ethpandaops/benchmarkoor/pkg/executor"
 	"github.com/ethpandaops/benchmarkoor/pkg/fsutil"
-	"github.com/ethpandaops/benchmarkoor/pkg/jsonrpc"
-	"github.com/ethpandaops/benchmarkoor/pkg/podman"
-	"github.com/ethpandaops/benchmarkoor/pkg/stats"
 	"github.com/ethpandaops/benchmarkoor/pkg/upload"
-	"github.com/shirou/gopsutil/v4/cpu"
-	"github.com/shirou/gopsutil/v4/host"
-	"github.com/shirou/gopsutil/v4/mem"
 	"github.com/sirupsen/logrus"
 )
 
@@ -538,2375 +522,6 @@ func (r *runner) RunInstance(ctx context.Context, instance *config.ClientInstanc
 	)
 }
 
-// runContainerLifecycle runs a single container lifecycle: load genesis,
-// create container, start, wait for RPC, execute tests, stop.
-//
-//nolint:gocognit,cyclop // Container lifecycle is inherently complex.
-func (r *runner) runContainerLifecycle(
-	ctx context.Context,
-	params *containerRunParams,
-	spec client.Spec,
-	datadirCfg *config.DataDirConfig,
-	useDataDir bool,
-) error {
-	instance := params.Instance
-	runID := params.RunID
-	runResultsDir := params.RunResultsDir
-	benchmarkoorLogFile := params.BenchmarkoorLog
-	genesisSource := params.GenesisSource
-
-	log := r.log.WithFields(logrus.Fields{
-		"instance": instance.ID,
-		"run_id":   runID,
-	})
-
-	if params.GenesisGroupHash != "" {
-		log = log.WithField("genesis_group", params.GenesisGroupHash)
-	}
-
-	// Each container lifecycle manages its own cleanup and crash detection.
-	var localCleanupFuncs []func()
-
-	localCleanupStarted := make(chan struct{})
-
-	var localCleanupOnce sync.Once
-
-	defer func() {
-		localCleanupOnce.Do(func() { close(localCleanupStarted) })
-
-		for i := len(localCleanupFuncs) - 1; i >= 0; i-- {
-			localCleanupFuncs[i]()
-		}
-	}()
-
-	// Setup data directory: either container volume or copied datadir.
-	// Each container lifecycle gets a fresh volume/datadir.
-	var dataMount docker.Mount
-
-	if useDataDir {
-		log.WithFields(logrus.Fields{
-			"source": datadirCfg.SourceDir,
-			"method": datadirCfg.Method,
-		}).Info("Using pre-populated data directory")
-
-		provider, err := datadir.NewProvider(log, datadirCfg.Method)
-		if err != nil {
-			return fmt.Errorf("creating datadir provider: %w", err)
-		}
-
-		prepared, err := provider.Prepare(ctx, &datadir.ProviderConfig{
-			SourceDir:  datadirCfg.SourceDir,
-			InstanceID: instance.ID,
-			TmpDir:     r.cfg.TmpDataDir,
-		})
-		if err != nil {
-			return fmt.Errorf("preparing datadir: %w", err)
-		}
-
-		localCleanupFuncs = append(localCleanupFuncs, func() {
-			if cleanupErr := prepared.Cleanup(); cleanupErr != nil {
-				log.WithError(cleanupErr).Warn("Failed to cleanup datadir")
-			}
-		})
-
-		containerDir := datadirCfg.ContainerDir
-		if containerDir == "" {
-			containerDir = spec.DataDir()
-		}
-
-		dataMount = docker.Mount{
-			Type:   "bind",
-			Source: prepared.MountPath,
-			Target: containerDir,
-		}
-	} else if r.cfg.FullConfig != nil &&
-		r.cfg.FullConfig.GetRollbackStrategy(instance) == config.RollbackStrategyCheckpointRestore {
-		// Checkpoint-restore without a pre-populated datadir uses a bind
-		// mount to a host temp directory so the copy-based rollback
-		// manager can snapshot and rsync the data between tests.
-		tmpDir, mkErr := os.MkdirTemp("", fmt.Sprintf("benchmarkoor-cpdata-%s-", instance.ID))
-		if mkErr != nil {
-			return fmt.Errorf("creating temp dir for checkpoint-restore data: %w", mkErr)
-		}
-
-		localCleanupFuncs = append(localCleanupFuncs, func() {
-			if rmErr := os.RemoveAll(tmpDir); rmErr != nil {
-				log.WithError(rmErr).Warn("Failed to remove checkpoint data temp dir")
-			}
-		})
-
-		log.WithField("path", tmpDir).Info(
-			"Using bind mount for checkpoint-restore (no datadir)",
-		)
-
-		dataMount = docker.Mount{
-			Type:   "bind",
-			Source: tmpDir,
-			Target: spec.DataDir(),
-		}
-	} else {
-		volumeSuffix := instance.ID
-		if params.GenesisGroupHash != "" {
-			volumeSuffix = instance.ID + "-" + params.GenesisGroupHash
-		}
-
-		volumeName := fmt.Sprintf("benchmarkoor-%s-%s", runID, volumeSuffix)
-		volumeLabels := map[string]string{
-			"benchmarkoor.instance":   instance.ID,
-			"benchmarkoor.client":     instance.Client,
-			"benchmarkoor.run-id":     runID,
-			"benchmarkoor.managed-by": "benchmarkoor",
-		}
-
-		if err := r.containerMgr.CreateVolume(
-			ctx, volumeName, volumeLabels,
-		); err != nil {
-			return fmt.Errorf("creating volume: %w", err)
-		}
-
-		localCleanupFuncs = append(localCleanupFuncs, func() {
-			if rmErr := r.containerMgr.RemoveVolume(
-				context.Background(), volumeName,
-			); rmErr != nil {
-				log.WithError(rmErr).Warn("Failed to remove volume")
-			}
-		})
-
-		dataMount = docker.Mount{
-			Type:   "volume",
-			Source: volumeName,
-			Target: spec.DataDir(),
-		}
-	}
-
-	// Load genesis file if configured.
-	var genesisContent []byte
-
-	if genesisSource != "" {
-		log.WithField("source", genesisSource).Info("Loading genesis file")
-
-		var loadErr error
-
-		genesisContent, loadErr = r.loadFile(ctx, genesisSource)
-		if loadErr != nil {
-			return fmt.Errorf("loading genesis: %w", loadErr)
-		}
-	} else {
-		log.Info("No genesis configured, skipping genesis setup")
-	}
-
-	// Fail if neither genesis nor datadir is configured.
-	if genesisSource == "" && !useDataDir {
-		return fmt.Errorf(
-			"no genesis file or datadir configured for client %s",
-			instance.Client,
-		)
-	}
-
-	// Image is already pulled by RunInstance; use the resolved name and digest.
-	imageName := params.ImageName
-	imageDigest := params.ImageDigest
-
-	// Create temp files for genesis and JWT.
-	tempDir, err := os.MkdirTemp(
-		r.cfg.TmpCacheDir, "benchmarkoor-"+instance.ID+"-",
-	)
-	if err != nil {
-		return fmt.Errorf("creating temp directory: %w", err)
-	}
-
-	localCleanupFuncs = append(localCleanupFuncs, func() {
-		if rmErr := os.RemoveAll(tempDir); rmErr != nil {
-			log.WithError(rmErr).Warn("Failed to remove temp directory")
-		}
-	})
-
-	// Write genesis file to temp dir if genesis is configured.
-	var genesisFile string
-
-	if genesisSource != "" {
-		genesisFile = filepath.Join(tempDir, "genesis.json")
-		if err := os.WriteFile(genesisFile, genesisContent, 0644); err != nil {
-			return fmt.Errorf("writing genesis file: %w", err)
-		}
-	}
-
-	jwtFile := filepath.Join(tempDir, "jwtsecret")
-	if err := os.WriteFile(jwtFile, []byte(r.cfg.JWT), 0644); err != nil {
-		return fmt.Errorf("writing jwt file: %w", err)
-	}
-
-	// Build container mounts.
-	mounts := []docker.Mount{
-		dataMount,
-		{
-			Type:     "bind",
-			Source:   jwtFile,
-			Target:   spec.JWTPath(),
-			ReadOnly: true,
-		},
-	}
-
-	// Add genesis mount if genesis is configured.
-	if genesisFile != "" {
-		mounts = append(mounts, docker.Mount{
-			Type:     "bind",
-			Source:   genesisFile,
-			Target:   spec.GenesisPath(),
-			ReadOnly: true,
-		})
-	}
-
-	// Run init container if required (skip when using datadir or no genesis).
-	if spec.RequiresInit() && !useDataDir && genesisSource != "" {
-		log.Info("Running init container")
-
-		initSuffix := "init"
-		if params.GenesisGroupHash != "" {
-			initSuffix = "init-" + params.GenesisGroupHash
-		}
-
-		initSpec := &docker.ContainerSpec{
-			Name: fmt.Sprintf(
-				"benchmarkoor-%s-%s-%s", runID, instance.ID, initSuffix,
-			),
-			Image:       imageName,
-			Command:     spec.InitCommand(),
-			Mounts:      mounts,
-			NetworkName: r.cfg.DockerNetwork,
-			Labels: map[string]string{
-				"benchmarkoor.instance":   instance.ID,
-				"benchmarkoor.client":     instance.Client,
-				"benchmarkoor.run-id":     runID,
-				"benchmarkoor.type":       "init",
-				"benchmarkoor.managed-by": "benchmarkoor",
-			},
-		}
-
-		// Set up init container log streaming (appends to container.log).
-		initLogFile := filepath.Join(runResultsDir, "container.log")
-
-		initFile, err := os.OpenFile(
-			initLogFile, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0644,
-		)
-		if err != nil {
-			return fmt.Errorf("opening init log file: %w", err)
-		}
-
-		if r.cfg.ResultsOwner != nil {
-			fsutil.Chown(initLogFile, r.cfg.ResultsOwner)
-		}
-
-		_, _ = fmt.Fprint(initFile, formatStartMarker("INIT_CONTAINER", &containerLogInfo{
-			Name:             initSpec.Name,
-			Image:            initSpec.Image,
-			GenesisGroupHash: params.GenesisGroupHash,
-		}))
-
-		var initStdout, initStderr io.Writer = initFile, initFile
-		if r.cfg.ClientLogsToStdout {
-			pfxFn := clientLogPrefix(instance.ID + "-init")
-			stdoutPrefixWriter := &prefixedWriter{
-				prefixFn: pfxFn, writer: os.Stdout,
-			}
-			logFilePrefixWriter := &prefixedWriter{
-				prefixFn: pfxFn, writer: benchmarkoorLogFile,
-			}
-			initStdout = io.MultiWriter(
-				initFile, stdoutPrefixWriter, logFilePrefixWriter,
-			)
-			initStderr = io.MultiWriter(
-				initFile, stdoutPrefixWriter, logFilePrefixWriter,
-			)
-		}
-
-		if err := r.containerMgr.RunInitContainer(
-			ctx, initSpec, initStdout, initStderr,
-		); err != nil {
-			_, _ = fmt.Fprintf(initFile, "#INIT_CONTAINER:END\n")
-			_ = initFile.Close()
-
-			return fmt.Errorf("running init container: %w", err)
-		}
-
-		_, _ = fmt.Fprintf(initFile, "#INIT_CONTAINER:END\n")
-		_ = initFile.Close()
-
-		log.Info("Init container completed")
-	} else if spec.RequiresInit() && genesisSource == "" {
-		log.Info("Skipping init container (no genesis configured)")
-	} else if useDataDir {
-		log.Info("Skipping init container (using pre-populated datadir)")
-	}
-
-	// Determine command.
-	cmd := make([]string, len(instance.Command))
-	copy(cmd, instance.Command)
-
-	if len(cmd) == 0 {
-		cmd = spec.DefaultCommand()
-	}
-
-	// Add genesis flag if genesis is configured and client uses a genesis flag.
-	if genesisSource != "" && spec.GenesisFlag() != "" {
-		cmd = append(cmd, spec.GenesisFlag()+spec.GenesisPath())
-	}
-
-	// Append extra args if provided, replacing any base args that share a flag prefix.
-	if len(instance.ExtraArgs) > 0 {
-		// Build set of flag prefixes from extra_args (e.g. "--config=" from "--config=mainnet.cfg").
-		prefixes := make([]string, 0, len(instance.ExtraArgs))
-		for _, arg := range instance.ExtraArgs {
-			if idx := strings.Index(arg, "="); idx != -1 {
-				prefixes = append(prefixes, arg[:idx+1])
-			}
-		}
-
-		// Remove any existing args that share a prefix with an extra arg.
-		if len(prefixes) > 0 {
-			filtered := make([]string, 0, len(cmd))
-			for _, c := range cmd {
-				override := false
-				for _, p := range prefixes {
-					if strings.HasPrefix(c, p) {
-						override = true
-
-						break
-					}
-				}
-
-				if !override {
-					filtered = append(filtered, c)
-				}
-			}
-
-			cmd = filtered
-		}
-
-		cmd = append(cmd, instance.ExtraArgs...)
-	}
-
-	// Build environment (default first, instance overrides).
-	env := make(
-		map[string]string,
-		len(spec.DefaultEnvironment())+len(instance.Environment),
-	)
-	for k, v := range spec.DefaultEnvironment() {
-		env[k] = v
-	}
-
-	for k, v := range instance.Environment {
-		env[k] = v
-	}
-
-	// When using checkpoint-restore, apply CRIU-compatibility env overrides.
-	if r.cfg.FullConfig != nil &&
-		r.cfg.FullConfig.GetRollbackStrategy(instance) == config.RollbackStrategyCheckpointRestore {
-		// Disable MPTCP so CRIU can checkpoint TCP sockets. CRIU does not
-		// support MPTCP (protocol 262) and recent Go versions enable it
-		// by default.
-		if _, ok := env["GODEBUG"]; !ok {
-			env["GODEBUG"] = "multipathtcp=0"
-		}
-
-		if client.ClientType(instance.Client) == client.ClientNethermind {
-			// NLog's autoReload uses FileSystemWatcher (inotify) on the
-			// overlay rootfs. CRIU cannot dump inotify watches on overlayfs
-			// (open_by_handle_at fails). Extract NLog.config from the image,
-			// patch autoReload to false, and bind-mount the patched copy so
-			// NLog never creates the FileSystemWatcher.
-			cpMgr, ok := r.containerMgr.(podman.CheckpointManager)
-			if ok {
-				nlogContent, nlogErr := cpMgr.ReadFileFromImage(
-					ctx, imageName, "/nethermind/NLog.config",
-				)
-				if nlogErr != nil {
-					log.WithError(nlogErr).Warn(
-						"Failed to extract NLog.config from image",
-					)
-				} else {
-					patched := strings.Replace(
-						string(nlogContent),
-						`autoReload="true"`,
-						`autoReload="false"`,
-						1,
-					)
-
-					nlogFile := filepath.Join(tempDir, "NLog.config")
-					if writeErr := os.WriteFile(
-						nlogFile, []byte(patched), 0644,
-					); writeErr != nil {
-						log.WithError(writeErr).Warn(
-							"Failed to write patched NLog.config",
-						)
-					} else {
-						mounts = append(mounts, docker.Mount{
-							Type:     "bind",
-							Source:   nlogFile,
-							Target:   "/nethermind/NLog.config",
-							ReadOnly: true,
-						})
-
-						log.Info(
-							"Bind-mounted patched NLog.config " +
-								"(autoReload=false) for CRIU compatibility",
-						)
-					}
-				}
-			}
-		}
-	}
-
-	// Resolve drop_memory_caches setting.
-	var dropMemoryCaches string
-	if r.cfg.FullConfig != nil {
-		dropMemoryCaches = r.cfg.FullConfig.GetDropMemoryCaches(instance)
-	}
-
-	// Resolve resource limits.
-	var containerResourceLimits *docker.ResourceLimits
-	var resolvedResourceLimits *ResolvedResourceLimits
-	var targetCPUs []int // CPUs to apply cpu_freq settings to
-
-	if r.cfg.FullConfig != nil {
-		resourceLimitsCfg := r.cfg.FullConfig.GetResourceLimits(instance)
-		if resourceLimitsCfg != nil {
-			var err error
-
-			containerResourceLimits, resolvedResourceLimits, err =
-				buildContainerResourceLimits(resourceLimitsCfg)
-			if err != nil {
-				return fmt.Errorf("building resource limits: %w", err)
-			}
-
-			fields := logrus.Fields{
-				"cpuset_cpus":   resolvedResourceLimits.CpusetCpus,
-				"memory":        resolvedResourceLimits.Memory,
-				"swap_disabled": resolvedResourceLimits.SwapDisabled,
-			}
-
-			if resolvedResourceLimits.BlkioConfig != nil {
-				fields["blkio_read_bps_devices"] = len(resolvedResourceLimits.BlkioConfig.DeviceReadBps)
-				fields["blkio_write_bps_devices"] = len(resolvedResourceLimits.BlkioConfig.DeviceWriteBps)
-				fields["blkio_read_iops_devices"] = len(resolvedResourceLimits.BlkioConfig.DeviceReadIOps)
-				fields["blkio_write_iops_devices"] = len(resolvedResourceLimits.BlkioConfig.DeviceWriteIOps)
-			}
-
-			log.WithFields(fields).Info("Resource limits configured")
-
-			// Determine target CPUs for cpu_freq settings.
-			// Use the resolved cpuset if available.
-			if resolvedResourceLimits.CpusetCpus != "" {
-				for _, cpuStr := range strings.Split(resolvedResourceLimits.CpusetCpus, ",") {
-					if cpuID, err := strconv.Atoi(strings.TrimSpace(cpuStr)); err == nil {
-						targetCPUs = append(targetCPUs, cpuID)
-					}
-				}
-			}
-
-			// Apply CPU frequency settings if configured.
-			if r.cpufreqMgr != nil && hasCPUFreqSettings(resourceLimitsCfg) {
-				cpufreqCfg := buildCPUFreqConfig(resourceLimitsCfg)
-
-				if err := r.cpufreqMgr.Apply(ctx, cpufreqCfg, targetCPUs); err != nil {
-					return fmt.Errorf("applying CPU frequency settings: %w", err)
-				}
-
-				// Log CPU frequency info.
-				logCPUFreqInfo(log, r.cpufreqMgr, targetCPUs)
-
-				// Add restore to cleanup.
-				localCleanupFuncs = append(localCleanupFuncs, func() {
-					if restoreErr := r.cpufreqMgr.Restore(context.Background()); restoreErr != nil {
-						log.WithError(restoreErr).Warn("Failed to restore CPU frequency settings")
-					}
-				})
-
-				// Update resolved limits with CPU freq info.
-				if cpufreqCfg.Frequency != "" && strings.ToUpper(cpufreqCfg.Frequency) != "MAX" {
-					if freqKHz, err := cpufreq.ParseFrequency(cpufreqCfg.Frequency); err == nil {
-						resolvedResourceLimits.CPUFreqKHz = &freqKHz
-					}
-				}
-				resolvedResourceLimits.CPUTurboBoost = cpufreqCfg.TurboBoost
-				resolvedResourceLimits.CPUGovernor = cpufreqCfg.Governor
-			}
-		}
-	}
-
-	// Resolve wait_after_rpc_ready for config output.
-	var waitAfterRPCReadyStr string
-	if r.cfg.FullConfig != nil {
-		if d := r.cfg.FullConfig.GetWaitAfterRPCReady(instance); d > 0 {
-			waitAfterRPCReadyStr = d.String()
-		}
-	}
-
-	// Resolve run_timeout for config output and timeout enforcement.
-	var runTimeoutStr string
-	var runTimeout time.Duration
-	if r.cfg.FullConfig != nil {
-		runTimeout = r.cfg.FullConfig.GetRunTimeout(instance)
-		if runTimeout > 0 {
-			runTimeoutStr = runTimeout.String()
-		}
-	}
-
-	// Write run configuration with resolved values.
-	runConfig := &RunConfig{
-		Timestamp: params.RunTimestamp,
-		System:    getSystemInfo(),
-		Instance: &ResolvedInstance{
-			ID:     instance.ID,
-			Client: instance.Client,
-			ContainerRuntime: func() string {
-				if r.cfg.FullConfig != nil {
-					return r.cfg.FullConfig.GetContainerRuntime()
-				}
-				return "docker"
-			}(),
-			Image:       imageName,
-			ImageSHA256: imageDigest,
-			Entrypoint:  instance.Entrypoint,
-			Command:     cmd,
-			ExtraArgs:   instance.ExtraArgs,
-			PullPolicy:  instance.PullPolicy,
-			Restart:     instance.Restart,
-			Environment: env,
-			DataDir:     datadirCfg,
-			RollbackStrategy: func() string {
-				if r.cfg.FullConfig != nil {
-					return r.cfg.FullConfig.GetRollbackStrategy(instance)
-				}
-				return ""
-			}(),
-			DropMemoryCaches:  dropMemoryCaches,
-			WaitAfterRPCReady: waitAfterRPCReadyStr,
-			RunTimeout:        runTimeoutStr,
-			RetryNewPayloadsSyncingState: func() *config.RetryNewPayloadsSyncingConfig {
-				if r.cfg.FullConfig != nil {
-					return r.cfg.FullConfig.GetRetryNewPayloadsSyncingState(instance)
-				}
-				return nil
-			}(),
-			ResourceLimits: resolvedResourceLimits,
-			PostTestRPCCalls: func() []config.PostTestRPCCall {
-				if r.cfg.FullConfig != nil {
-					return r.cfg.FullConfig.GetPostTestRPCCalls(instance)
-				}
-				return nil
-			}(),
-			BootstrapFCU: func() *config.BootstrapFCUConfig {
-				if r.cfg.FullConfig != nil {
-					return r.cfg.FullConfig.GetBootstrapFCU(instance)
-				}
-				return nil
-			}(),
-			CheckpointRestoreStrategyOptions: func() *config.CheckpointRestoreStrategyOptions {
-				if r.cfg.FullConfig != nil {
-					return r.cfg.FullConfig.GetCheckpointRestoreStrategyOptions(instance)
-				}
-				return nil
-			}(),
-		},
-	}
-
-	// Attach metadata labels if configured.
-	if r.cfg.FullConfig != nil && len(r.cfg.FullConfig.Runner.Metadata.Labels) > 0 {
-		runConfig.Metadata = &r.cfg.FullConfig.Runner.Metadata
-	}
-
-	if len(params.GenesisGroups) > 0 {
-		runConfig.Instance.GenesisGroups = params.GenesisGroups
-	} else {
-		runConfig.Instance.Genesis = genesisSource
-	}
-
-	if r.executor != nil {
-		runConfig.SuiteHash = r.executor.GetSuiteHash()
-	}
-
-	if err := writeRunConfig(
-		runResultsDir, runConfig, r.cfg.ResultsOwner,
-	); err != nil {
-		log.WithError(err).Warn("Failed to write run config")
-	}
-
-	// Build container spec.
-	containerName := fmt.Sprintf("benchmarkoor-%s-%s", runID, instance.ID)
-	if params.GenesisGroupHash != "" {
-		containerName = fmt.Sprintf(
-			"benchmarkoor-%s-%s-%s",
-			runID, instance.ID, params.GenesisGroupHash,
-		)
-	}
-
-	containerSpec := &docker.ContainerSpec{
-		Name:           containerName,
-		Image:          imageName,
-		Entrypoint:     instance.Entrypoint,
-		Command:        cmd,
-		Env:            env,
-		Mounts:         mounts,
-		NetworkName:    r.cfg.DockerNetwork,
-		ResourceLimits: containerResourceLimits,
-		Labels: map[string]string{
-			"benchmarkoor.instance":   instance.ID,
-			"benchmarkoor.client":     instance.Client,
-			"benchmarkoor.run-id":     runID,
-			"benchmarkoor.managed-by": "benchmarkoor",
-		},
-	}
-
-	// Save container spec and datadir info for runner-level rollback strategies.
-	params.ContainerSpec = containerSpec
-	params.DataDirCfg = datadirCfg
-	params.UseDataDir = useDataDir
-
-	// Create container.
-	containerID, err := r.containerMgr.CreateContainer(ctx, containerSpec)
-	if err != nil {
-		return fmt.Errorf("creating container: %w", err)
-	}
-
-	// Setup log streaming.
-	logCtx, logCancel := context.WithCancel(ctx)
-
-	logFilePath := filepath.Join(runResultsDir, "container.log")
-
-	logFile, err := os.OpenFile(logFilePath, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0644)
-	if err != nil {
-		logCancel()
-
-		return fmt.Errorf("opening container log file: %w", err)
-	}
-
-	if r.cfg.ResultsOwner != nil {
-		fsutil.Chown(logFilePath, r.cfg.ResultsOwner)
-	}
-
-	// Create block log collector to capture JSON payloads from client logs.
-	blockLogParser := blocklog.NewParser(client.ClientType(instance.Client))
-	blockLogCollector := blocklog.NewCollector(blockLogParser, logFile)
-	params.BlockLogCollector = blockLogCollector
-
-	logDone := make(chan struct{})
-
-	r.wg.Add(1)
-
-	go func() {
-		defer r.wg.Done()
-		defer close(logDone)
-
-		if err := r.streamLogs(
-			logCtx, instance.ID, containerID, logFile, benchmarkoorLogFile,
-			&containerLogInfo{
-				Name:             containerName,
-				ContainerID:      containerID,
-				Image:            imageName,
-				GenesisGroupHash: params.GenesisGroupHash,
-			},
-			blockLogCollector,
-		); err != nil {
-			// Context cancellation during cleanup is expected.
-			select {
-			case <-localCleanupStarted:
-				log.WithError(err).Debug("Log streaming stopped")
-			default:
-				log.WithError(err).Warn("Log streaming error")
-			}
-		}
-	}()
-
-	// Cleanup funcs execute in LIFO order. Add the log-drain + file-close
-	// first, then the container removal. At teardown the container is removed
-	// (stopping it → Docker flushes remaining logs → StreamLogs returns EOF),
-	// then waitForLogDrain ensures the goroutine has finished writing before
-	// the log file is closed.
-	localCleanupFuncs = append(localCleanupFuncs, func() {
-		waitForLogDrain(&logDone, &logCancel, logDrainTimeout)
-		_ = logFile.Close()
-	})
-
-	localCleanupFuncs = append(localCleanupFuncs, func() {
-		log.Info("Removing container")
-
-		if rmErr := r.containerMgr.RemoveContainer(
-			context.Background(), containerID,
-		); rmErr != nil {
-			log.WithError(rmErr).Warn("Failed to remove container")
-		}
-	})
-
-	// Start container.
-	if err := r.containerMgr.StartContainer(ctx, containerID); err != nil {
-		return fmt.Errorf("starting container: %w", err)
-	}
-
-	log.Info("Container started")
-
-	// Apply run timeout if configured.
-	testCtx := ctx
-	var timeoutCancel context.CancelFunc
-
-	if runTimeout > 0 {
-		testCtx, timeoutCancel = context.WithTimeout(ctx, runTimeout)
-		defer timeoutCancel()
-
-		log.WithField("timeout", runTimeout).Info("Run timeout configured")
-	}
-
-	// Start container death monitoring.
-	execCtx, execCancel := context.WithCancel(testCtx)
-	defer execCancel()
-
-	var containerDied bool
-	var containerExitCode *int64
-	var containerOOMKilled *bool
-	var mu sync.Mutex
-
-	containerExitCh, containerErrCh := r.containerMgr.WaitForContainerExit(
-		ctx, containerID,
-	)
-
-	r.wg.Add(1)
-
-	go func() {
-		defer r.wg.Done()
-
-		select {
-		case exitInfo := <-containerExitCh:
-			mu.Lock()
-			containerDied = true
-			containerExitCode = &exitInfo.ExitCode
-			containerOOMKilled = &exitInfo.OOMKilled
-			mu.Unlock()
-
-			logFields := logrus.Fields{
-				"exit_code":  exitInfo.ExitCode,
-				"oom_killed": exitInfo.OOMKilled,
-			}
-
-			select {
-			case <-localCleanupStarted:
-				log.WithFields(logFields).Debug(
-					"Container stopped during cleanup",
-				)
-			default:
-				log.WithFields(logFields).Warn(
-					"Container exited unexpectedly",
-				)
-			}
-
-			execCancel()
-		case err := <-containerErrCh:
-			if err != nil && !errors.Is(err, context.Canceled) {
-				log.WithError(err).Warn("Container wait error")
-			}
-		case <-r.done:
-			// Runner is stopping.
-		}
-	}()
-
-	// Get container IP for health checks.
-	containerIP, err := r.containerMgr.GetContainerIP(
-		ctx, containerID, r.cfg.DockerNetwork,
-	)
-	if err != nil {
-		return fmt.Errorf("getting container IP: %w", err)
-	}
-
-	log.WithField("ip", containerIP).Debug("Container IP address")
-
-	// Wait for RPC to be ready.
-	clientVersion, err := r.waitForRPC(execCtx, containerIP, spec.RPCPort())
-	if err != nil {
-		mu.Lock()
-		if containerDied {
-			runConfig.Status = RunStatusContainerDied
-			runConfig.TerminationReason = fmt.Sprintf(
-				"container exited while waiting for RPC: %v", err,
-			)
-			runConfig.ContainerExitCode = containerExitCode
-			runConfig.ContainerOOMKilled = containerOOMKilled
-		} else {
-			runConfig.Status = RunStatusFailed
-			runConfig.TerminationReason = fmt.Sprintf(
-				"waiting for RPC: %v", err,
-			)
-		}
-		runConfig.TimestampEnd = time.Now().Unix()
-		mu.Unlock()
-
-		if writeErr := writeRunConfig(
-			runResultsDir, runConfig, r.cfg.ResultsOwner,
-		); writeErr != nil {
-			log.WithError(writeErr).Warn(
-				"Failed to write run config with failed status",
-			)
-		}
-
-		return fmt.Errorf("waiting for RPC: %w", err)
-	}
-
-	log.WithField("version", clientVersion).Info("RPC endpoint ready")
-
-	// Wait after RPC ready if configured (gives client time to complete internal sync).
-	if r.cfg.FullConfig != nil {
-		if waitDuration := r.cfg.FullConfig.GetWaitAfterRPCReady(instance); waitDuration > 0 {
-			log.WithField("duration", waitDuration).Info("Waiting after RPC ready")
-
-			select {
-			case <-time.After(waitDuration):
-			case <-execCtx.Done():
-				return execCtx.Err()
-			}
-		}
-	}
-
-	// Log the latest block info.
-	blockNum, blockHash, stateRoot, blkErr := r.getLatestBlock(execCtx, containerIP, spec.RPCPort())
-	if blkErr != nil {
-		log.WithError(blkErr).Warn("Failed to get latest block")
-	} else {
-		log.WithFields(logrus.Fields{
-			"block_number": blockNum,
-			"block_hash":   blockHash,
-			"state_root":   stateRoot,
-		}).Info("Latest block")
-
-		runConfig.StartBlock = &StartBlock{
-			Number:    blockNum,
-			Hash:      blockHash,
-			StateRoot: stateRoot,
-		}
-	}
-
-	// Send bootstrap FCU if configured.
-	if r.cfg.FullConfig != nil {
-		if fcuCfg := r.cfg.FullConfig.GetBootstrapFCU(instance); fcuCfg != nil && fcuCfg.Enabled {
-			fcuHash := blockHash
-			if fcuCfg.HeadBlockHash != "" {
-				fcuHash = fcuCfg.HeadBlockHash
-
-				log.WithField("block_hash", fcuHash).Info(
-					"Using configured block hash for bootstrap FCU",
-				)
-			}
-
-			if fcuHash != "" {
-				if fcuErr := r.sendBootstrapFCU(
-					execCtx, log, containerIP, spec.EnginePort(), fcuHash, fcuCfg,
-				); fcuErr != nil {
-					log.WithError(fcuErr).Error("Bootstrap FCU failed")
-
-					return fmt.Errorf("sending bootstrap FCU: %w", fcuErr)
-				}
-
-				// Re-fetch latest block after FCU with a configured head_block_hash
-				// so that runConfig.StartBlock reflects the post-FCU state.
-				if fcuCfg.HeadBlockHash != "" {
-					bn, bh, sr, err := r.getLatestBlock(execCtx, containerIP, spec.RPCPort())
-					if err != nil {
-						log.WithError(err).Warn("Failed to get latest block after bootstrap FCU")
-					} else {
-						log.WithFields(logrus.Fields{
-							"block_number": bn,
-							"block_hash":   bh,
-							"state_root":   sr,
-						}).Info("Latest block after bootstrap FCU")
-
-						runConfig.StartBlock = &StartBlock{
-							Number:    bn,
-							Hash:      bh,
-							StateRoot: sr,
-						}
-					}
-				}
-			}
-		}
-	}
-
-	// Update config with client version.
-	runConfig.Instance.ClientVersion = clientVersion
-
-	if err := writeRunConfig(
-		runResultsDir, runConfig, r.cfg.ResultsOwner,
-	); err != nil {
-		log.WithError(err).Warn(
-			"Failed to update run config with client version",
-		)
-	}
-
-	// Execute tests if executor is configured.
-	if r.executor != nil {
-		log.Info("Starting test execution")
-
-		var dropCachesPath string
-		if r.cfg.FullConfig != nil {
-			dropCachesPath = r.cfg.FullConfig.GetDropCachesPath()
-		}
-
-		// Resolve rollback strategy.
-		var rollbackStrategy string
-		if r.cfg.FullConfig != nil {
-			rollbackStrategy = r.cfg.FullConfig.GetRollbackStrategy(instance)
-		}
-
-		// Fail fast if checkpoint-restore prerequisites are not met.
-		if rollbackStrategy == config.RollbackStrategyCheckpointRestore {
-			cpMgr, ok := r.containerMgr.(podman.CheckpointManager)
-			if !ok {
-				return fmt.Errorf("container manager does not support checkpoint/restore")
-			}
-
-			if err := cpMgr.ValidateCheckpointSupport(ctx); err != nil {
-				return fmt.Errorf("checkpoint/restore prerequisites not met: %w", err)
-			}
-		}
-
-		isRunnerLevel := rollbackStrategy == config.RollbackStrategyContainerRecreate ||
-			rollbackStrategy == config.RollbackStrategyCheckpointRestore
-
-		var (
-			result  *executor.ExecutionResult
-			execErr error
-		)
-
-		if isRunnerLevel {
-			// Runner-level strategies intentionally stop and restart
-			// containers. Signal cleanup-started so the death monitor
-			// treats container exits as expected (debug-level logging),
-			// and cancel execCtx so the monitor's execCancel() is a no-op.
-			localCleanupOnce.Do(func() { close(localCleanupStarted) })
-			execCancel()
-
-			switch rollbackStrategy {
-			case config.RollbackStrategyCheckpointRestore:
-				result, execErr = r.runTestsWithCheckpointRestore(
-					testCtx, params, spec, containerID, containerIP,
-					dropMemoryCaches, dropCachesPath,
-					runResultsDir, &logCancel, &logDone, benchmarkoorLogFile,
-					&localCleanupFuncs, localCleanupStarted,
-				)
-			default:
-				result, execErr = r.runTestsWithContainerStrategy(
-					testCtx, params, spec, containerID, containerIP,
-					rollbackStrategy, dropMemoryCaches, dropCachesPath,
-					runResultsDir, &logCancel, &logDone, benchmarkoorLogFile,
-					&localCleanupFuncs, localCleanupStarted,
-				)
-			}
-		} else {
-			execOpts := &executor.ExecuteOptions{
-				EngineEndpoint: fmt.Sprintf(
-					"http://%s:%d", containerIP, spec.EnginePort(),
-				),
-				JWT:                   r.cfg.JWT,
-				ResultsDir:            runResultsDir,
-				Filter:                r.cfg.TestFilter,
-				ContainerID:           containerID,
-				DockerClient:          r.getDockerClient(),
-				DropMemoryCaches:      dropMemoryCaches,
-				DropCachesPath:        dropCachesPath,
-				RollbackStrategy:      rollbackStrategy,
-				ClientRPCRollbackSpec: spec.RPCRollbackSpec(),
-				RPCEndpoint: fmt.Sprintf(
-					"http://%s:%d", containerIP, spec.RPCPort(),
-				),
-				Tests:                         params.Tests,
-				BlockLogCollector:             params.BlockLogCollector,
-				RetryNewPayloadsSyncingConfig: r.cfg.FullConfig.GetRetryNewPayloadsSyncingState(instance),
-				PostTestRPCCalls:              r.cfg.FullConfig.GetPostTestRPCCalls(instance),
-			}
-
-			result, execErr = r.executor.ExecuteTests(execCtx, execOpts)
-		}
-
-		if execErr != nil {
-			log.WithError(execErr).Error("Test execution failed")
-
-			mu.Lock()
-			runConfig.Status = RunStatusFailed
-			runConfig.TerminationReason = fmt.Sprintf(
-				"test execution failed: %v", execErr,
-			)
-			mu.Unlock()
-		}
-
-		if result != nil {
-			log.WithFields(logrus.Fields{
-				"total":    result.TotalTests,
-				"passed":   result.Passed,
-				"failed":   result.Failed,
-				"duration": result.TotalDuration,
-			}).Info("Test execution completed")
-
-			suiteTotal := result.TotalTests
-			if r.executor != nil {
-				if allTests := r.executor.GetTests(); allTests != nil {
-					suiteTotal = len(allTests)
-				}
-			}
-
-			mu.Lock()
-
-			if params.AccumulatedTestCount != nil {
-				// Multi-genesis mode: accumulate counts across groups.
-				params.AccumulatedTestCount.Total = suiteTotal
-				params.AccumulatedTestCount.Passed += result.Passed
-				params.AccumulatedTestCount.Failed += result.Failed
-				runConfig.TestCounts = &TestCounts{
-					Total:  params.AccumulatedTestCount.Total,
-					Passed: params.AccumulatedTestCount.Passed,
-					Failed: params.AccumulatedTestCount.Failed,
-				}
-			} else {
-				runConfig.TestCounts = &TestCounts{
-					Total:  suiteTotal,
-					Passed: result.Passed,
-					Failed: result.Failed,
-				}
-			}
-
-			mu.Unlock()
-
-			if result.StatsReaderType != "" {
-				runConfig.SystemResourceCollectionMethod = result.StatsReaderType
-			}
-
-			if isRunnerLevel {
-				// Runner-level strategies intentionally stop containers,
-				// which causes the death monitor to set containerDied.
-				// Reset it and trust only the strategy's result.
-				mu.Lock()
-				containerDied = result.ContainerDied
-				containerExitCode = nil
-				containerOOMKilled = nil
-				mu.Unlock()
-			} else if result.ContainerDied {
-				mu.Lock()
-				containerDied = true
-				mu.Unlock()
-			}
-		}
-	}
-
-	// Determine final run status (don't overwrite if already set by executor).
-	// Timeout and cancellation are checked before containerDied because when
-	// either fires, the context cancellation stops the container, which causes
-	// the death monitor to set containerDied.
-	mu.Lock()
-	if timeoutCancel != nil && testCtx.Err() == context.DeadlineExceeded {
-		runConfig.Status = RunStatusTimedOut
-		runConfig.TerminationReason = fmt.Sprintf("the run_timeout of %s was reached", runTimeout)
-	} else if ctx.Err() != nil {
-		runConfig.Status = RunStatusCancelled
-		runConfig.TerminationReason = "run was cancelled"
-	} else if containerDied {
-		runConfig.Status = RunStatusContainerDied
-		runConfig.TerminationReason = "container exited during test execution"
-		runConfig.ContainerExitCode = containerExitCode
-		runConfig.ContainerOOMKilled = containerOOMKilled
-	} else if runConfig.Status == "" {
-		runConfig.Status = RunStatusCompleted
-	}
-	mu.Unlock()
-
-	// Record when the run ended.
-	runConfig.TimestampEnd = time.Now().Unix()
-
-	// Write final config with status.
-	if err := writeRunConfig(
-		runResultsDir, runConfig, r.cfg.ResultsOwner,
-	); err != nil {
-		log.WithError(err).Warn("Failed to write final run config with status")
-	} else {
-		log.WithField("status", runConfig.Status).Info("Run completed")
-	}
-
-	// Write block logs if any were captured.
-	if params.BlockLogCollector != nil {
-		blockLogs := params.BlockLogCollector.GetBlockLogs()
-		if len(blockLogs) > 0 {
-			if err := executor.WriteBlockLogsResult(
-				runResultsDir, blockLogs, r.cfg.ResultsOwner,
-			); err != nil {
-				log.WithError(err).Warn("Failed to write block logs result")
-			} else {
-				log.WithField("count", len(blockLogs)).Info("Block logs written")
-			}
-		}
-	}
-
-	// Return an error if the container died so callers (e.g. multi-genesis
-	// loop) stop instead of continuing with the next group.
-	if containerDied {
-		return fmt.Errorf("container died during execution")
-	}
-
-	return nil
-}
-
-// runTestsWithContainerStrategy executes tests one at a time, manipulating
-// the container between tests according to the given strategy.
-//
-//nolint:gocognit,cyclop // Per-test container manipulation is inherently complex.
-func (r *runner) runTestsWithContainerStrategy(
-	ctx context.Context,
-	params *containerRunParams,
-	spec client.Spec,
-	containerID string,
-	containerIP string,
-	strategy string,
-	dropMemoryCaches string,
-	dropCachesPath string,
-	resultsDir string,
-	logCancel *context.CancelFunc,
-	logDone *chan struct{},
-	benchmarkoorLog *os.File,
-	cleanupFuncs *[]func(),
-	cleanupStarted chan struct{},
-) (*executor.ExecutionResult, error) {
-	log := r.log.WithFields(logrus.Fields{
-		"instance": params.Instance.ID,
-		"run_id":   params.RunID,
-		"strategy": strategy,
-	})
-
-	// Resolve the test list.
-	tests := params.Tests
-	if tests == nil {
-		tests = r.executor.GetTests()
-	}
-
-	if len(tests) == 0 {
-		return &executor.ExecutionResult{}, nil
-	}
-
-	log.WithField("tests", len(tests)).Info(
-		"Running tests with container-level rollback strategy",
-	)
-
-	// Detect ZFS snapshot optimization: when the strategy is
-	// container-recreate with a ZFS datadir, we snapshot the data
-	// directory after the first RPC-ready state, then rollback to
-	// that snapshot between tests instead of cloning from scratch.
-	useZFSSnapshot := strategy == config.RollbackStrategyContainerRecreate &&
-		params.DataDirCfg != nil && params.DataDirCfg.Method == "zfs"
-
-	// snapshotRollback holds the rollback/cleanup callbacks for the
-	// ZFS snapshot path. Only populated when useZFSSnapshot is true.
-	type snapshotRollback struct {
-		rollback func(ctx context.Context) error
-		cleanup  func()
-	}
-
-	var sr *snapshotRollback
-
-	if useZFSSnapshot {
-		log.Info(
-			"ZFS datadir detected: will snapshot after RPC-ready " +
-				"and rollback between tests",
-		)
-
-		// Stop the initial container so writes are flushed to disk.
-		if err := r.containerMgr.StopContainer(ctx, containerID); err != nil {
-			return nil, fmt.Errorf("stopping container for ZFS snapshot: %w", err)
-		}
-
-		waitForLogDrain(logDone, logCancel, logDrainTimeout)
-
-		// Sync to flush any dirty pages before snapshotting.
-		if syncErr := exec.Command("sync").Run(); syncErr != nil {
-			log.WithError(syncErr).Warn(
-				"Failed to sync before ZFS snapshot",
-			)
-		}
-
-		// Find the data mount source path from the container spec.
-		containerDir := spec.DataDir()
-		if params.DataDirCfg.ContainerDir != "" {
-			containerDir = params.DataDirCfg.ContainerDir
-		}
-
-		dataMountSource := ""
-
-		for _, mnt := range params.ContainerSpec.Mounts {
-			if mnt.Target == containerDir {
-				dataMountSource = mnt.Source
-
-				break
-			}
-		}
-
-		if dataMountSource == "" {
-			return nil, fmt.Errorf(
-				"could not find data mount for %s in container spec",
-				containerDir,
-			)
-		}
-
-		// Take the ready-state ZFS snapshot.
-		zfsMgr := datadir.NewCheckpointZFSManager(r.log)
-
-		snapshot, snapErr := zfsMgr.SnapshotReady(
-			ctx, &datadir.CheckpointConfig{
-				DataDir:    dataMountSource,
-				InstanceID: params.Instance.ID,
-			},
-		)
-		if snapErr != nil {
-			return nil, fmt.Errorf(
-				"creating ready-state ZFS snapshot: %w", snapErr,
-			)
-		}
-
-		sr = &snapshotRollback{
-			rollback: func(ctx context.Context) error {
-				return zfsMgr.RollbackToReady(ctx, snapshot)
-			},
-			cleanup: func() {
-				if destroyErr := zfsMgr.DestroySnapshot(snapshot); destroyErr != nil {
-					log.WithError(destroyErr).Warn(
-						"Failed to destroy ready-state ZFS snapshot",
-					)
-				}
-			},
-		}
-
-		defer sr.cleanup()
-
-		// Remove the initial container (we'll create fresh ones per test).
-		if err := r.containerMgr.RemoveContainer(
-			ctx, containerID,
-		); err != nil {
-			log.WithError(err).Warn("Failed to remove initial container")
-		}
-	}
-
-	combined := &executor.ExecutionResult{}
-	startTime := time.Now()
-	currentContainerID := containerID
-	currentContainerIP := containerIP
-
-	for i, test := range tests {
-		select {
-		case <-ctx.Done():
-			waitForLogDrain(logDone, logCancel, logDrainTimeout)
-			combined.TotalDuration = time.Since(startTime)
-
-			return combined, ctx.Err()
-		default:
-		}
-
-		testLog := log.WithFields(logrus.Fields{
-			"test":  test.Name,
-			"index": fmt.Sprintf("%d/%d", i+1, len(tests)),
-		})
-
-		// Restore state before test.
-		switch {
-		case useZFSSnapshot:
-			// ZFS snapshot path: rollback datadir, create a fresh
-			// container on the same mount, start, wait for RPC.
-			testLog.Info("Rolling back ZFS snapshot for next test")
-
-			if i > 0 {
-				// Stop and remove container from previous test.
-				if err := r.containerMgr.StopContainer(
-					ctx, currentContainerID,
-				); err != nil {
-					testLog.WithError(err).Warn(
-						"Failed to stop container",
-					)
-				}
-
-				waitForLogDrain(logDone, logCancel, logDrainTimeout)
-
-				if err := r.containerMgr.RemoveContainer(
-					ctx, currentContainerID,
-				); err != nil {
-					testLog.WithError(err).Warn(
-						"Failed to remove container",
-					)
-				}
-			}
-
-			// Rollback the ZFS dataset to the ready-state snapshot.
-			if err := sr.rollback(ctx); err != nil {
-				combined.TotalDuration = time.Since(startTime)
-
-				return combined, fmt.Errorf(
-					"rolling back ZFS snapshot for test %d: %w", i, err,
-				)
-			}
-
-			// Create a new container using the same mount path.
-			newSpec := *params.ContainerSpec
-			newSpec.Name = fmt.Sprintf("%s-%d", params.ContainerSpec.Name, i)
-			newSpec.Mounts = make(
-				[]docker.Mount, len(params.ContainerSpec.Mounts),
-			)
-			copy(newSpec.Mounts, params.ContainerSpec.Mounts)
-
-			newID, err := r.containerMgr.CreateContainer(ctx, &newSpec)
-			if err != nil {
-				combined.TotalDuration = time.Since(startTime)
-
-				return combined, fmt.Errorf(
-					"creating container for test %d: %w", i, err,
-				)
-			}
-
-			currentContainerID = newID
-
-			*cleanupFuncs = append(*cleanupFuncs, func() {
-				if rmErr := r.containerMgr.RemoveContainer(
-					context.Background(), newID,
-				); rmErr != nil {
-					testLog.WithError(rmErr).Warn(
-						"Failed to remove recreated container",
-					)
-				}
-			})
-
-			// Start fresh log streaming.
-			if err := r.startLogStreaming(
-				ctx, resultsDir,
-				params.Instance.ID, newID,
-				benchmarkoorLog, &containerLogInfo{
-					Name:             newSpec.Name,
-					ContainerID:      newID,
-					Image:            newSpec.Image,
-					GenesisGroupHash: params.GenesisGroupHash,
-				},
-				params.BlockLogCollector, cleanupStarted,
-				logDone, logCancel, cleanupFuncs,
-			); err != nil {
-				combined.TotalDuration = time.Since(startTime)
-
-				return combined, fmt.Errorf(
-					"log streaming for test %d: %w", i, err,
-				)
-			}
-
-			// Start the new container.
-			if err := r.containerMgr.StartContainer(ctx, newID); err != nil {
-				waitForLogDrain(logDone, logCancel, logDrainTimeout)
-				combined.TotalDuration = time.Since(startTime)
-
-				return combined, fmt.Errorf(
-					"starting container for test %d: %w", i, err,
-				)
-			}
-
-			// Get new container IP.
-			newIP, err := r.containerMgr.GetContainerIP(
-				ctx, newID, r.cfg.DockerNetwork,
-			)
-			if err != nil {
-				waitForLogDrain(logDone, logCancel, logDrainTimeout)
-				combined.TotalDuration = time.Since(startTime)
-
-				return combined, fmt.Errorf(
-					"getting container IP for test %d: %w", i, err,
-				)
-			}
-
-			currentContainerIP = newIP
-
-			// Wait for RPC to be ready.
-			if _, err := r.waitForRPC(
-				ctx, currentContainerIP, spec.RPCPort(),
-			); err != nil {
-				waitForLogDrain(logDone, logCancel, logDrainTimeout)
-				combined.TotalDuration = time.Since(startTime)
-
-				return combined, fmt.Errorf(
-					"waiting for RPC on test %d: %w", i, err,
-				)
-			}
-
-			// Send bootstrap FCU if configured.
-			if r.cfg.FullConfig != nil {
-				if fcuCfg := r.cfg.FullConfig.GetBootstrapFCU(
-					params.Instance,
-				); fcuCfg != nil && fcuCfg.Enabled {
-					blkHash := fcuCfg.HeadBlockHash
-					if blkHash == "" {
-						var blkErr error
-						_, blkHash, _, blkErr = r.getLatestBlock(
-							ctx, currentContainerIP, spec.RPCPort(),
-						)
-
-						if blkErr != nil {
-							testLog.WithError(blkErr).Warn(
-								"Failed to get latest block " +
-									"for bootstrap FCU",
-							)
-						}
-					}
-
-					if blkHash != "" {
-						if fcuErr := r.sendBootstrapFCU(
-							ctx, testLog, currentContainerIP,
-							spec.EnginePort(), blkHash, fcuCfg,
-						); fcuErr != nil {
-							testLog.WithError(fcuErr).Error(
-								"Bootstrap FCU failed",
-							)
-							waitForLogDrain(
-								logDone, logCancel, logDrainTimeout,
-							)
-							combined.TotalDuration = time.Since(startTime)
-
-							return combined, fmt.Errorf(
-								"sending bootstrap FCU for test %d: %w",
-								i, fcuErr,
-							)
-						}
-					}
-				}
-			}
-
-		case strategy == config.RollbackStrategyContainerRecreate && i > 0:
-			testLog.Info("Recreating container for next test")
-
-			// Stop container first so Docker flushes remaining logs.
-			if err := r.containerMgr.StopContainer(
-				ctx, currentContainerID,
-			); err != nil {
-				testLog.WithError(err).Warn("Failed to stop container")
-			}
-
-			// Wait for the log-streaming goroutine to finish.
-			waitForLogDrain(logDone, logCancel, logDrainTimeout)
-
-			// Remove the stopped container.
-			if err := r.containerMgr.RemoveContainer(
-				ctx, currentContainerID,
-			); err != nil {
-				testLog.WithError(err).Warn("Failed to remove container")
-			}
-
-			// Create a fresh data volume/datadir for the new container.
-			newSpec := *params.ContainerSpec
-			newSpec.Name = fmt.Sprintf("%s-%d", params.ContainerSpec.Name, i)
-			newSpec.Mounts = make([]docker.Mount, len(params.ContainerSpec.Mounts))
-			copy(newSpec.Mounts, params.ContainerSpec.Mounts)
-
-			freshMount, mountCleanup, err := r.createFreshDataMount(
-				ctx, params, spec, i,
-			)
-			if err != nil {
-				combined.TotalDuration = time.Since(startTime)
-
-				return combined, fmt.Errorf(
-					"creating fresh data mount for test %d: %w", i, err,
-				)
-			}
-
-			if mountCleanup != nil {
-				*cleanupFuncs = append(*cleanupFuncs, mountCleanup)
-			}
-
-			// Replace the data mount (index 0) with the fresh one.
-			newSpec.Mounts[0] = freshMount
-
-			// Run init container if required to populate the fresh volume.
-			if spec.RequiresInit() && !params.UseDataDir &&
-				params.GenesisSource != "" {
-				testLog.Info("Running init container for fresh volume")
-
-				initMounts := make([]docker.Mount, len(newSpec.Mounts))
-				copy(initMounts, newSpec.Mounts)
-
-				if err := r.runInitForRecreate(
-					ctx, params, spec, initMounts, resultsDir,
-					benchmarkoorLog, i,
-				); err != nil {
-					combined.TotalDuration = time.Since(startTime)
-
-					return combined, fmt.Errorf(
-						"running init container for test %d: %w", i, err,
-					)
-				}
-			}
-
-			newID, err := r.containerMgr.CreateContainer(ctx, &newSpec)
-			if err != nil {
-				combined.TotalDuration = time.Since(startTime)
-
-				return combined, fmt.Errorf("creating container for test %d: %w", i, err)
-			}
-
-			currentContainerID = newID
-
-			// Update cleanup to remove this container on exit.
-			*cleanupFuncs = append(*cleanupFuncs, func() {
-				if rmErr := r.containerMgr.RemoveContainer(
-					context.Background(), newID,
-				); rmErr != nil {
-					testLog.WithError(rmErr).Warn(
-						"Failed to remove recreated container",
-					)
-				}
-			})
-
-			// Start fresh log streaming.
-			if err := r.startLogStreaming(
-				ctx, resultsDir,
-				params.Instance.ID, newID,
-				benchmarkoorLog, &containerLogInfo{
-					Name:             newSpec.Name,
-					ContainerID:      newID,
-					Image:            newSpec.Image,
-					GenesisGroupHash: params.GenesisGroupHash,
-				},
-				params.BlockLogCollector, cleanupStarted,
-				logDone, logCancel, cleanupFuncs,
-			); err != nil {
-				combined.TotalDuration = time.Since(startTime)
-
-				return combined, fmt.Errorf(
-					"log streaming for test %d: %w", i, err,
-				)
-			}
-
-			// Start the new container.
-			if err := r.containerMgr.StartContainer(ctx, newID); err != nil {
-				waitForLogDrain(logDone, logCancel, logDrainTimeout)
-				combined.TotalDuration = time.Since(startTime)
-
-				return combined, fmt.Errorf("starting container for test %d: %w", i, err)
-			}
-
-			// Get new container IP.
-			newIP, err := r.containerMgr.GetContainerIP(
-				ctx, newID, r.cfg.DockerNetwork,
-			)
-			if err != nil {
-				waitForLogDrain(logDone, logCancel, logDrainTimeout)
-				combined.TotalDuration = time.Since(startTime)
-
-				return combined, fmt.Errorf("getting container IP for test %d: %w", i, err)
-			}
-
-			currentContainerIP = newIP
-
-			// Wait for RPC to be ready.
-			if _, err := r.waitForRPC(
-				ctx, currentContainerIP, spec.RPCPort(),
-			); err != nil {
-				waitForLogDrain(logDone, logCancel, logDrainTimeout)
-				combined.TotalDuration = time.Since(startTime)
-
-				return combined, fmt.Errorf("waiting for RPC on test %d: %w", i, err)
-			}
-
-			// Send bootstrap FCU if configured.
-			if r.cfg.FullConfig != nil {
-				if fcuCfg := r.cfg.FullConfig.GetBootstrapFCU(params.Instance); fcuCfg != nil && fcuCfg.Enabled {
-					blkHash := fcuCfg.HeadBlockHash
-					if blkHash == "" {
-						var blkErr error
-						_, blkHash, _, blkErr = r.getLatestBlock(
-							ctx, currentContainerIP, spec.RPCPort(),
-						)
-
-						if blkErr != nil {
-							testLog.WithError(blkErr).Warn(
-								"Failed to get latest block for bootstrap FCU",
-							)
-						}
-					}
-
-					if blkHash != "" {
-						if fcuErr := r.sendBootstrapFCU(
-							ctx, testLog, currentContainerIP,
-							spec.EnginePort(), blkHash, fcuCfg,
-						); fcuErr != nil {
-							testLog.WithError(fcuErr).Error(
-								"Bootstrap FCU failed",
-							)
-							waitForLogDrain(logDone, logCancel, logDrainTimeout)
-							combined.TotalDuration = time.Since(startTime)
-
-							return combined, fmt.Errorf(
-								"sending bootstrap FCU for test %d: %w",
-								i, fcuErr,
-							)
-						}
-					}
-				}
-			}
-
-		}
-
-		testLog.Info("Executing test")
-
-		// Execute single test via executor with no executor-level rollback.
-		execOpts := &executor.ExecuteOptions{
-			EngineEndpoint: fmt.Sprintf(
-				"http://%s:%d", currentContainerIP, spec.EnginePort(),
-			),
-			JWT:              r.cfg.JWT,
-			ResultsDir:       resultsDir,
-			Filter:           r.cfg.TestFilter,
-			ContainerID:      currentContainerID,
-			DockerClient:     r.getDockerClient(),
-			DropMemoryCaches: dropMemoryCaches,
-			DropCachesPath:   dropCachesPath,
-			RollbackStrategy: config.RollbackStrategyNone,
-			RPCEndpoint: fmt.Sprintf(
-				"http://%s:%d", currentContainerIP, spec.RPCPort(),
-			),
-			Tests:                         []*executor.TestWithSteps{test},
-			BlockLogCollector:             params.BlockLogCollector,
-			RetryNewPayloadsSyncingConfig: r.cfg.FullConfig.GetRetryNewPayloadsSyncingState(params.Instance),
-			PostTestRPCCalls:              r.cfg.FullConfig.GetPostTestRPCCalls(params.Instance),
-		}
-
-		result, err := r.executor.ExecuteTests(ctx, execOpts)
-		if err != nil {
-			testLog.WithError(err).Error("Test execution failed")
-
-			continue
-		}
-
-		// Aggregate results.
-		combined.TotalTests += result.TotalTests
-		combined.Passed += result.Passed
-		combined.Failed += result.Failed
-
-		if result.StatsReaderType != "" {
-			combined.StatsReaderType = result.StatsReaderType
-		}
-
-		if result.ContainerDied {
-			combined.ContainerDied = true
-			combined.TotalDuration = time.Since(startTime)
-
-			waitForLogDrain(logDone, logCancel, logDrainTimeout)
-
-			return combined, nil
-		}
-	}
-
-	combined.TotalDuration = time.Since(startTime)
-
-	waitForLogDrain(logDone, logCancel, logDrainTimeout)
-
-	return combined, nil
-}
-
-// createFreshDataMount creates a new volume or datadir for a recreated container.
-// Returns the mount, a cleanup function (may be nil), and any error.
-func (r *runner) createFreshDataMount(
-	ctx context.Context,
-	params *containerRunParams,
-	spec client.Spec,
-	iteration int,
-) (docker.Mount, func(), error) {
-	log := r.log.WithFields(logrus.Fields{
-		"instance":  params.Instance.ID,
-		"run_id":    params.RunID,
-		"iteration": iteration,
-	})
-
-	if params.UseDataDir {
-		log.Info("Preparing fresh datadir copy")
-
-		provider, err := datadir.NewProvider(log, params.DataDirCfg.Method)
-		if err != nil {
-			return docker.Mount{}, nil, fmt.Errorf("creating datadir provider: %w", err)
-		}
-
-		prepared, err := provider.Prepare(ctx, &datadir.ProviderConfig{
-			SourceDir:  params.DataDirCfg.SourceDir,
-			InstanceID: fmt.Sprintf("%s-%d", params.Instance.ID, iteration),
-			TmpDir:     r.cfg.TmpDataDir,
-		})
-		if err != nil {
-			return docker.Mount{}, nil, fmt.Errorf("preparing datadir: %w", err)
-		}
-
-		containerDir := params.DataDirCfg.ContainerDir
-		if containerDir == "" {
-			containerDir = spec.DataDir()
-		}
-
-		cleanup := func() {
-			if cleanupErr := prepared.Cleanup(); cleanupErr != nil {
-				log.WithError(cleanupErr).Warn("Failed to cleanup recreate datadir")
-			}
-		}
-
-		return docker.Mount{
-			Type:   "bind",
-			Source: prepared.MountPath,
-			Target: containerDir,
-		}, cleanup, nil
-	}
-
-	// Container volume path.
-	volumeSuffix := params.Instance.ID
-	if params.GenesisGroupHash != "" {
-		volumeSuffix = params.Instance.ID + "-" + params.GenesisGroupHash
-	}
-
-	volumeName := fmt.Sprintf(
-		"benchmarkoor-%s-%s-%d", params.RunID, volumeSuffix, iteration,
-	)
-	volumeLabels := map[string]string{
-		"benchmarkoor.instance":   params.Instance.ID,
-		"benchmarkoor.client":     params.Instance.Client,
-		"benchmarkoor.run-id":     params.RunID,
-		"benchmarkoor.managed-by": "benchmarkoor",
-	}
-
-	if err := r.containerMgr.CreateVolume(ctx, volumeName, volumeLabels); err != nil {
-		return docker.Mount{}, nil, fmt.Errorf("creating volume: %w", err)
-	}
-
-	log.WithField("volume", volumeName).Debug("Created fresh volume")
-
-	cleanup := func() {
-		if rmErr := r.containerMgr.RemoveVolume(
-			context.Background(), volumeName,
-		); rmErr != nil {
-			log.WithError(rmErr).Warn("Failed to remove recreate volume")
-		}
-	}
-
-	return docker.Mount{
-		Type:   "volume",
-		Source: volumeName,
-		Target: spec.DataDir(),
-	}, cleanup, nil
-}
-
-// runInitForRecreate runs an init container to populate a fresh volume
-// during container-recreate strategy.
-func (r *runner) runInitForRecreate(
-	ctx context.Context,
-	params *containerRunParams,
-	spec client.Spec,
-	mounts []docker.Mount,
-	resultsDir string,
-	benchmarkoorLog *os.File,
-	iteration int,
-) error {
-	instance := params.Instance
-
-	initName := fmt.Sprintf(
-		"benchmarkoor-%s-%s-init-%d", params.RunID, instance.ID, iteration,
-	)
-	if params.GenesisGroupHash != "" {
-		initName = fmt.Sprintf(
-			"benchmarkoor-%s-%s-%s-init-%d",
-			params.RunID, instance.ID, params.GenesisGroupHash, iteration,
-		)
-	}
-
-	initSpec := &docker.ContainerSpec{
-		Name:        initName,
-		Image:       params.ImageName,
-		Command:     spec.InitCommand(),
-		Mounts:      mounts,
-		NetworkName: r.cfg.DockerNetwork,
-		Labels: map[string]string{
-			"benchmarkoor.instance":   instance.ID,
-			"benchmarkoor.client":     instance.Client,
-			"benchmarkoor.run-id":     params.RunID,
-			"benchmarkoor.type":       "init",
-			"benchmarkoor.managed-by": "benchmarkoor",
-		},
-	}
-
-	initLogFile := filepath.Join(resultsDir, "container.log")
-
-	initFile, err := os.OpenFile(
-		initLogFile, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0644,
-	)
-	if err != nil {
-		return fmt.Errorf("opening init log file: %w", err)
-	}
-
-	if r.cfg.ResultsOwner != nil {
-		fsutil.Chown(initLogFile, r.cfg.ResultsOwner)
-	}
-
-	_, _ = fmt.Fprint(initFile, formatStartMarker("INIT_CONTAINER", &containerLogInfo{
-		Name:             initSpec.Name,
-		Image:            initSpec.Image,
-		GenesisGroupHash: params.GenesisGroupHash,
-	}))
-
-	var initStdout, initStderr io.Writer = initFile, initFile
-	if r.cfg.ClientLogsToStdout {
-		pfxFn := clientLogPrefix(instance.ID + "-init")
-		stdoutPW := &prefixedWriter{prefixFn: pfxFn, writer: os.Stdout}
-		logPW := &prefixedWriter{prefixFn: pfxFn, writer: benchmarkoorLog}
-		initStdout = io.MultiWriter(initFile, stdoutPW, logPW)
-		initStderr = io.MultiWriter(initFile, stdoutPW, logPW)
-	}
-
-	if err := r.containerMgr.RunInitContainer(
-		ctx, initSpec, initStdout, initStderr,
-	); err != nil {
-		_, _ = fmt.Fprintf(initFile, "#INIT_CONTAINER:END\n")
-		_ = initFile.Close()
-
-		return fmt.Errorf("running init container: %w", err)
-	}
-
-	_, _ = fmt.Fprintf(initFile, "#INIT_CONTAINER:END\n")
-	_ = initFile.Close()
-
-	return nil
-}
-
-// loadFile loads content from a URL or local file path.
-func (r *runner) loadFile(ctx context.Context, source string) ([]byte, error) {
-	// Check if source is a URL.
-	if strings.HasPrefix(source, "http://") || strings.HasPrefix(source, "https://") {
-		return r.downloadFromURL(ctx, source)
-	}
-
-	// Treat as local file path.
-	return r.readFromFile(source)
-}
-
-// downloadFromURL downloads content from a URL.
-func (r *runner) downloadFromURL(ctx context.Context, url string) ([]byte, error) {
-	req, err := http.NewRequestWithContext(ctx, http.MethodGet, url, nil)
-	if err != nil {
-		return nil, fmt.Errorf("creating request: %w", err)
-	}
-
-	resp, err := http.DefaultClient.Do(req)
-	if err != nil {
-		return nil, fmt.Errorf("executing request: %w", err)
-	}
-	defer func() { _ = resp.Body.Close() }()
-
-	if resp.StatusCode != http.StatusOK {
-		return nil, fmt.Errorf("unexpected status code: %d", resp.StatusCode)
-	}
-
-	data, err := io.ReadAll(resp.Body)
-	if err != nil {
-		return nil, fmt.Errorf("reading response body: %w", err)
-	}
-
-	return data, nil
-}
-
-// readFromFile reads content from a local file.
-func (r *runner) readFromFile(path string) ([]byte, error) {
-	data, err := os.ReadFile(path)
-	if err != nil {
-		return nil, fmt.Errorf("reading file %s: %w", path, err)
-	}
-
-	return data, nil
-}
-
-// getSystemInfo gathers system hardware and OS information.
-func getSystemInfo() *SystemInfo {
-	info := &SystemInfo{}
-
-	if hostInfo, err := host.Info(); err == nil {
-		info.Hostname = hostInfo.Hostname
-		info.OS = hostInfo.OS
-		info.Platform = hostInfo.Platform
-		info.PlatformVersion = hostInfo.PlatformVersion
-		info.KernelVersion = hostInfo.KernelVersion
-		info.Arch = hostInfo.KernelArch
-		info.Virtualization = hostInfo.VirtualizationSystem
-		info.VirtualizationRole = hostInfo.VirtualizationRole
-	}
-
-	if cpuInfo, err := cpu.Info(); err == nil && len(cpuInfo) > 0 {
-		info.CPUVendor = cpuInfo[0].VendorID
-		info.CPUModel = cpuInfo[0].ModelName
-		info.CPUMhz = cpuInfo[0].Mhz
-		info.CPUCacheKB = int(cpuInfo[0].CacheSize)
-	}
-
-	if cores, err := cpu.Counts(false); err == nil {
-		info.CPUCores = cores
-	}
-
-	if memInfo, err := mem.VirtualMemory(); err == nil {
-		info.MemoryTotalGB = float64(memInfo.Total) / (1024 * 1024 * 1024)
-	}
-
-	return info
-}
-
-// writeRunConfig writes the run configuration to config.json.
-func writeRunConfig(resultsDir string, cfg *RunConfig, owner *fsutil.OwnerConfig) error {
-	data, err := json.MarshalIndent(cfg, "", "  ")
-	if err != nil {
-		return fmt.Errorf("marshaling run config: %w", err)
-	}
-
-	configPath := filepath.Join(resultsDir, "config.json")
-	if err := fsutil.WriteFile(configPath, data, 0644, owner); err != nil {
-		return fmt.Errorf("writing config.json: %w", err)
-	}
-
-	return nil
-}
-
-// streamLogs streams container logs to file and optionally stdout/benchmarkoor log.
-// The log file should be opened in append mode before calling this function.
-// If blockLogCollector is provided, the collector's writer wraps the file writer
-// to intercept and parse JSON payloads from log lines.
-func (r *runner) streamLogs(
-	ctx context.Context,
-	instanceID, containerID string,
-	file *os.File,
-	benchmarkoorLog io.Writer,
-	logInfo *containerLogInfo,
-	blockLogCollector blocklog.Collector,
-) error {
-	// Write start marker with container metadata.
-	_, _ = fmt.Fprint(file, formatStartMarker("CONTAINER", logInfo))
-
-	// Base writer is the file, optionally wrapped by block log collector.
-	var baseWriter io.Writer = file
-	if blockLogCollector != nil {
-		baseWriter = blockLogCollector.Writer()
-	}
-
-	stdout, stderr := baseWriter, baseWriter
-
-	if r.cfg.ClientLogsToStdout {
-		pfxFn := clientLogPrefix(instanceID)
-		stdoutPrefixWriter := &prefixedWriter{prefixFn: pfxFn, writer: os.Stdout}
-		logFilePrefixWriter := &prefixedWriter{prefixFn: pfxFn, writer: benchmarkoorLog}
-		stdout = io.MultiWriter(baseWriter, stdoutPrefixWriter, logFilePrefixWriter)
-		stderr = io.MultiWriter(baseWriter, stdoutPrefixWriter, logFilePrefixWriter)
-	}
-
-	streamErr := r.containerMgr.StreamLogs(ctx, containerID, stdout, stderr)
-
-	// Write end marker (best-effort, even if streaming failed).
-	_, _ = fmt.Fprintf(file, "#CONTAINER:END\n")
-
-	return streamErr
-}
-
-// startLogStreaming opens the container log file in append mode, registers a
-// file-close cleanup, and launches a goroutine that streams container logs.
-// It writes through logDone/logCancel so the caller (and waitForLogDrain)
-// can drain and clean up the streaming goroutine.
-func (r *runner) startLogStreaming(
-	ctx context.Context,
-	resultsDir string,
-	instanceID, containerID string,
-	benchmarkoorLog io.Writer,
-	logInfo *containerLogInfo,
-	blockLogCollector blocklog.Collector,
-	cleanupStarted <-chan struct{},
-	logDone *chan struct{},
-	logCancel *context.CancelFunc,
-	cleanupFuncs *[]func(),
-) error {
-	logCtx, cancel := context.WithCancel(ctx)
-	*logCancel = cancel
-
-	logFilePath := filepath.Join(resultsDir, "container.log")
-
-	logFile, err := os.OpenFile(
-		logFilePath, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0644,
-	)
-	if err != nil {
-		cancel()
-
-		return fmt.Errorf("opening container log file: %w", err)
-	}
-
-	*cleanupFuncs = append(*cleanupFuncs, func() { _ = logFile.Close() })
-
-	done := make(chan struct{})
-	*logDone = done
-
-	r.wg.Add(1)
-
-	go func() {
-		defer r.wg.Done()
-		defer close(done)
-
-		if streamErr := r.streamLogs(
-			logCtx, instanceID, containerID, logFile,
-			benchmarkoorLog, logInfo, blockLogCollector,
-		); streamErr != nil {
-			select {
-			case <-cleanupStarted:
-			default:
-				r.log.WithError(streamErr).Debug("Log streaming ended")
-			}
-		}
-	}()
-
-	return nil
-}
-
-// waitForLogDrain waits for the log-streaming goroutine to finish (signalled
-// via logDone) up to the given timeout. If the timeout expires it cancels the
-// log context and waits for the goroutine to acknowledge. This must be called
-// *after* the container has been stopped so that Docker flushes buffered logs.
-func waitForLogDrain(
-	logDone *chan struct{},
-	logCancel *context.CancelFunc,
-	timeout time.Duration,
-) {
-	if logDone == nil || *logDone == nil {
-		if logCancel != nil {
-			(*logCancel)()
-		}
-
-		return
-	}
-
-	select {
-	case <-*logDone:
-	case <-time.After(timeout):
-		(*logCancel)()
-		<-*logDone
-	}
-}
-
-// waitForRPC waits for the RPC endpoint to be ready and returns the client version.
-func (r *runner) waitForRPC(ctx context.Context, host string, port int) (string, error) {
-	ctx, cancel := context.WithTimeout(ctx, r.cfg.ReadyTimeout)
-	defer cancel()
-
-	url := fmt.Sprintf("http://%s:%d", host, port)
-
-	ticker := time.NewTicker(DefaultHealthCheckInterval)
-	defer ticker.Stop()
-
-	for {
-		select {
-		case <-ctx.Done():
-			return "", fmt.Errorf("timeout waiting for RPC: %w", ctx.Err())
-		case <-ticker.C:
-			if version, ok := r.checkRPCHealth(ctx, url); ok {
-				return version, nil
-			}
-		}
-	}
-}
-
-// checkRPCHealth performs a single RPC health check and returns the client version on success.
-func (r *runner) checkRPCHealth(ctx context.Context, url string) (string, bool) {
-	ctx, cancel := context.WithTimeout(ctx, 5*time.Second)
-	defer cancel()
-
-	body := `{"jsonrpc":"2.0","method":"web3_clientVersion","params":[],"id":1}`
-
-	req, err := http.NewRequestWithContext(ctx, http.MethodPost, url, strings.NewReader(body))
-	if err != nil {
-		return "", false
-	}
-
-	req.Header.Set("Content-Type", "application/json")
-
-	resp, err := http.DefaultClient.Do(req)
-	if err != nil {
-		return "", false
-	}
-	defer func() { _ = resp.Body.Close() }()
-
-	if resp.StatusCode != http.StatusOK {
-		return "", false
-	}
-
-	respBody, err := io.ReadAll(resp.Body)
-	if err != nil {
-		return "", false
-	}
-
-	var rpcResp struct {
-		Result string `json:"result"`
-	}
-
-	if err := json.Unmarshal(respBody, &rpcResp); err != nil {
-		return "", false
-	}
-
-	return rpcResp.Result, true
-}
-
-// getLatestBlock fetches the latest block number, hash, and state root from the RPC endpoint.
-func (r *runner) getLatestBlock(ctx context.Context, host string, port int) (uint64, string, string, error) {
-	ctx, cancel := context.WithTimeout(ctx, 10*time.Second)
-	defer cancel()
-
-	url := fmt.Sprintf("http://%s:%d", host, port)
-	body := `{"jsonrpc":"2.0","method":"eth_getBlockByNumber","params":["latest",false],"id":1}`
-
-	req, err := http.NewRequestWithContext(ctx, http.MethodPost, url, strings.NewReader(body))
-	if err != nil {
-		return 0, "", "", fmt.Errorf("creating request: %w", err)
-	}
-
-	req.Header.Set("Content-Type", "application/json")
-
-	resp, err := http.DefaultClient.Do(req)
-	if err != nil {
-		return 0, "", "", fmt.Errorf("executing request: %w", err)
-	}
-	defer func() { _ = resp.Body.Close() }()
-
-	if resp.StatusCode != http.StatusOK {
-		return 0, "", "", fmt.Errorf("unexpected status code: %d", resp.StatusCode)
-	}
-
-	respBody, err := io.ReadAll(resp.Body)
-	if err != nil {
-		return 0, "", "", fmt.Errorf("reading response: %w", err)
-	}
-
-	var rpcResp struct {
-		Result struct {
-			Number    string `json:"number"`
-			Hash      string `json:"hash"`
-			StateRoot string `json:"stateRoot"`
-		} `json:"result"`
-	}
-
-	if err := json.Unmarshal(respBody, &rpcResp); err != nil {
-		return 0, "", "", fmt.Errorf("parsing response: %w", err)
-	}
-
-	// Parse hex block number.
-	blockNum, err := strconv.ParseUint(strings.TrimPrefix(rpcResp.Result.Number, "0x"), 16, 64)
-	if err != nil {
-		return 0, "", "", fmt.Errorf("parsing block number: %w", err)
-	}
-
-	return blockNum, rpcResp.Result.Hash, rpcResp.Result.StateRoot, nil
-}
-
-// sendBootstrapFCU sends an engine_forkchoiceUpdatedV3 call to confirm the
-// client is fully synced and ready for test execution. The call is retried
-// up to cfg.MaxRetries times with cfg.Backoff between attempts — some clients
-// (e.g., Erigon) may still be performing internal initialization after RPC
-// becomes available. A VALID response confirms the client is ready.
-func (r *runner) sendBootstrapFCU(
-	ctx context.Context,
-	log logrus.FieldLogger,
-	host string,
-	enginePort int,
-	headBlockHash string,
-	cfg *config.BootstrapFCUConfig,
-) error {
-	const zeroHash = "0x0000000000000000000000000000000000000000000000000000000000000000"
-
-	backoff, err := time.ParseDuration(cfg.Backoff)
-	if err != nil {
-		return fmt.Errorf("parsing backoff duration: %w", err)
-	}
-
-	// Build the forkchoiceUpdatedV3 payload.
-	payload := fmt.Sprintf(
-		`{"jsonrpc":"2.0","method":"engine_forkchoiceUpdatedV3",`+
-			`"params":[{"headBlockHash":"%s","safeBlockHash":"%s",`+
-			`"finalizedBlockHash":"%s"},null],"id":1}`,
-		headBlockHash, zeroHash, zeroHash,
-	)
-
-	url := fmt.Sprintf("http://%s:%d", host, enginePort)
-
-	log.WithFields(logrus.Fields{
-		"max_retries": cfg.MaxRetries,
-		"backoff":     cfg.Backoff,
-		"payload":     payload,
-	}).Info("Sending bootstrap FCU")
-
-	var lastErr error
-
-	for attempt := 1; attempt <= cfg.MaxRetries; attempt++ {
-		lastErr = r.doBootstrapFCURequest(ctx, url, payload)
-		if lastErr == nil {
-			log.WithField("head_block_hash", headBlockHash).Info(
-				"Bootstrap FCU sent successfully",
-			)
-
-			return nil
-		}
-
-		log.WithFields(logrus.Fields{
-			"attempt": attempt,
-			"max":     cfg.MaxRetries,
-			"error":   lastErr.Error(),
-		}).Warn("Bootstrap FCU attempt failed, retrying")
-
-		if attempt < cfg.MaxRetries {
-			select {
-			case <-time.After(backoff):
-			case <-ctx.Done():
-				return fmt.Errorf("context cancelled during retry: %w", ctx.Err())
-			}
-		}
-	}
-
-	return fmt.Errorf("bootstrap FCU failed after %d attempts: %w", cfg.MaxRetries, lastErr)
-}
-
-// doBootstrapFCURequest performs a single bootstrap FCU HTTP request.
-func (r *runner) doBootstrapFCURequest(
-	ctx context.Context,
-	url string,
-	payload string,
-) error {
-	const requestTimeout = 30 * time.Second
-
-	reqCtx, cancel := context.WithTimeout(ctx, requestTimeout)
-	defer cancel()
-
-	token, err := executor.GenerateJWTToken(r.cfg.JWT)
-	if err != nil {
-		return fmt.Errorf("generating JWT: %w", err)
-	}
-
-	req, err := http.NewRequestWithContext(
-		reqCtx, http.MethodPost, url, strings.NewReader(payload),
-	)
-	if err != nil {
-		return fmt.Errorf("creating request: %w", err)
-	}
-
-	req.Header.Set("Content-Type", "application/json")
-	req.Header.Set("Authorization", "Bearer "+token)
-
-	resp, err := http.DefaultClient.Do(req)
-	if err != nil {
-		return fmt.Errorf("executing request: %w", err)
-	}
-	defer func() { _ = resp.Body.Close() }()
-
-	if resp.StatusCode != http.StatusOK {
-		return fmt.Errorf("unexpected status code: %d", resp.StatusCode)
-	}
-
-	body, err := io.ReadAll(resp.Body)
-	if err != nil {
-		return fmt.Errorf("reading response: %w", err)
-	}
-
-	rpcResp, err := jsonrpc.Parse(string(body))
-	if err != nil {
-		return fmt.Errorf("parsing response: %w", err)
-	}
-
-	// Validate the response using the FCU validator.
-	validator := &jsonrpc.ForkchoiceUpdatedValidator{}
-	if err := validator.Validate("engine_forkchoiceUpdatedV3", rpcResp); err != nil {
-		return fmt.Errorf("validating response: %w", err)
-	}
-
-	return nil
-}
-
-// prefixedWriter adds a prefix to each line written.
-// If prefixFn is set, it is called per line to generate the prefix dynamically.
-// Otherwise, the static prefix field is used.
-type prefixedWriter struct {
-	prefix   string
-	prefixFn func() string
-	writer   io.Writer
-	buf      []byte
-}
-
-func (w *prefixedWriter) Write(p []byte) (n int, err error) {
-	n = len(p)
-	w.buf = append(w.buf, p...)
-
-	for {
-		idx := -1
-
-		for i, b := range w.buf {
-			if b == '\n' {
-				idx = i
-
-				break
-			}
-		}
-
-		if idx == -1 {
-			break
-		}
-
-		line := w.buf[:idx+1]
-		w.buf = w.buf[idx+1:]
-
-		pfx := w.prefix
-		if w.prefixFn != nil {
-			pfx = w.prefixFn()
-		}
-
-		if _, err := fmt.Fprintf(w.writer, "%s%s", pfx, line); err != nil {
-			return n, err
-		}
-	}
-
-	return n, nil
-}
-
-// clientLogPrefix returns a function that generates a consistent log prefix
-// for client container logs: "🟣 $TIMESTAMP CLIE | $clientName | ".
-func clientLogPrefix(clientName string) func() string {
-	return func() string {
-		ts := time.Now().UTC().Format(config.LogTimestampFormat)
-
-		return fmt.Sprintf("🟣 %s CLIE | %s | ", ts, clientName)
-	}
-}
-
-// fileHook writes log entries to a file.
-type fileHook struct {
-	writer    io.Writer
-	formatter logrus.Formatter
-}
-
-func (h *fileHook) Levels() []logrus.Level {
-	return logrus.AllLevels
-}
-
-func (h *fileHook) Fire(entry *logrus.Entry) error {
-	line, err := h.formatter.Format(entry)
-	if err != nil {
-		return err
-	}
-
-	_, err = h.writer.Write(line)
-
-	return err
-}
-
-// removeHook removes a hook from the logger.
-func (r *runner) removeHook(hook logrus.Hook) {
-	for level, hooks := range r.logger.Hooks {
-		filtered := make([]logrus.Hook, 0, len(hooks))
-
-		for _, h := range hooks {
-			if h != hook {
-				filtered = append(filtered, h)
-			}
-		}
-
-		r.logger.Hooks[level] = filtered
-	}
-}
-
 // generateShortID generates a short random hex ID (8 characters).
 func generateShortID() string {
 	b := make([]byte, 4)
@@ -2916,748 +531,4 @@ func generateShortID() string {
 	}
 
 	return hex.EncodeToString(b)
-}
-
-// selectRandomCPUs picks count random CPUs from available CPUs using Fisher-Yates shuffle.
-func selectRandomCPUs(count int) ([]int, error) {
-	numCPUs, err := cpu.Counts(true)
-	if err != nil {
-		return nil, fmt.Errorf("getting CPU count: %w", err)
-	}
-
-	if count > numCPUs {
-		return nil, fmt.Errorf("requested %d CPUs but only %d available", count, numCPUs)
-	}
-
-	// Create slice of all CPU IDs.
-	cpus := make([]int, numCPUs)
-	for i := range cpus {
-		cpus[i] = i
-	}
-
-	// Fisher-Yates shuffle (partial - only shuffle first 'count' elements).
-	for i := 0; i < count; i++ {
-		j := i + mrand.IntN(numCPUs-i)
-		cpus[i], cpus[j] = cpus[j], cpus[i]
-	}
-
-	return cpus[:count], nil
-}
-
-// cpusetString converts a slice of CPU IDs to a comma-separated string.
-func cpusetString(cpus []int) string {
-	if len(cpus) == 0 {
-		return ""
-	}
-
-	strs := make([]string, len(cpus))
-	for i, c := range cpus {
-		strs[i] = strconv.Itoa(c)
-	}
-
-	return strings.Join(strs, ",")
-}
-
-// buildContainerResourceLimits builds docker.ResourceLimits from config.ResourceLimits.
-func buildContainerResourceLimits(cfg *config.ResourceLimits) (*docker.ResourceLimits, *ResolvedResourceLimits, error) {
-	if cfg == nil {
-		return nil, nil, nil
-	}
-
-	containerLimits := &docker.ResourceLimits{}
-	resolved := &ResolvedResourceLimits{}
-
-	// Handle CPU pinning.
-	if cfg.CpusetCount != nil {
-		cpus, err := selectRandomCPUs(*cfg.CpusetCount)
-		if err != nil {
-			return nil, nil, fmt.Errorf("selecting random CPUs: %w", err)
-		}
-
-		containerLimits.CpusetCpus = cpusetString(cpus)
-		resolved.CpusetCpus = containerLimits.CpusetCpus
-	} else if len(cfg.Cpuset) > 0 {
-		containerLimits.CpusetCpus = cpusetString(cfg.Cpuset)
-		resolved.CpusetCpus = containerLimits.CpusetCpus
-	}
-
-	// Handle memory limit.
-	if cfg.Memory != "" {
-		memBytes, err := units.RAMInBytes(cfg.Memory)
-		if err != nil {
-			return nil, nil, fmt.Errorf("parsing memory limit: %w", err)
-		}
-
-		containerLimits.MemoryBytes = memBytes
-		resolved.Memory = cfg.Memory
-		resolved.MemoryBytes = memBytes
-
-		// Handle swap.
-		if cfg.SwapDisabled {
-			// Set memory-swap equal to memory to disable swap.
-			containerLimits.MemorySwapBytes = memBytes
-			// Set swappiness to 0.
-			swappiness := int64(0)
-			containerLimits.MemorySwappiness = &swappiness
-			resolved.SwapDisabled = true
-		}
-	}
-
-	// Handle blkio config.
-	if cfg.BlkioConfig != nil {
-		blkioCfg := cfg.BlkioConfig
-		resolvedBlkio := &ResolvedBlkioConfig{}
-
-		// Process device_read_bps.
-		if len(blkioCfg.DeviceReadBps) > 0 {
-			containerLimits.BlkioDeviceReadBps, resolvedBlkio.DeviceReadBps = convertBlkioDevicesBps(blkioCfg.DeviceReadBps)
-		}
-
-		// Process device_write_bps.
-		if len(blkioCfg.DeviceWriteBps) > 0 {
-			containerLimits.BlkioDeviceWriteBps, resolvedBlkio.DeviceWriteBps = convertBlkioDevicesBps(blkioCfg.DeviceWriteBps)
-		}
-
-		// Process device_read_iops.
-		if len(blkioCfg.DeviceReadIOps) > 0 {
-			containerLimits.BlkioDeviceReadIOps, resolvedBlkio.DeviceReadIOps = convertBlkioDevicesIOps(blkioCfg.DeviceReadIOps)
-		}
-
-		// Process device_write_iops.
-		if len(blkioCfg.DeviceWriteIOps) > 0 {
-			containerLimits.BlkioDeviceWriteIOps, resolvedBlkio.DeviceWriteIOps = convertBlkioDevicesIOps(blkioCfg.DeviceWriteIOps)
-		}
-
-		// Only set if we have any blkio config.
-		if len(resolvedBlkio.DeviceReadBps) > 0 || len(resolvedBlkio.DeviceWriteBps) > 0 ||
-			len(resolvedBlkio.DeviceReadIOps) > 0 || len(resolvedBlkio.DeviceWriteIOps) > 0 {
-			resolved.BlkioConfig = resolvedBlkio
-		}
-	}
-
-	return containerLimits, resolved, nil
-}
-
-// convertBlkioDevicesBps converts config blkio devices with bps rates to docker and resolved formats.
-func convertBlkioDevicesBps(devices []config.ThrottleDevice) ([]docker.BlkioThrottleDevice, []ResolvedThrottleDevice) {
-	dockerDevices := make([]docker.BlkioThrottleDevice, len(devices))
-	resolvedDevices := make([]ResolvedThrottleDevice, len(devices))
-
-	for i, dev := range devices {
-		// Parse rate using RAMInBytes (validation already done in config.Validate).
-		rate, _ := units.RAMInBytes(dev.Rate)
-
-		dockerDevices[i] = docker.BlkioThrottleDevice{
-			Path: dev.Path,
-			Rate: uint64(rate),
-		}
-		resolvedDevices[i] = ResolvedThrottleDevice{
-			Path: dev.Path,
-			Rate: uint64(rate),
-		}
-	}
-
-	return dockerDevices, resolvedDevices
-}
-
-// convertBlkioDevicesIOps converts config blkio devices with IOPS rates to docker and resolved formats.
-func convertBlkioDevicesIOps(devices []config.ThrottleDevice) ([]docker.BlkioThrottleDevice, []ResolvedThrottleDevice) {
-	dockerDevices := make([]docker.BlkioThrottleDevice, len(devices))
-	resolvedDevices := make([]ResolvedThrottleDevice, len(devices))
-
-	for i, dev := range devices {
-		// Parse rate as integer (validation already done in config.Validate).
-		rate, _ := strconv.ParseUint(dev.Rate, 10, 64)
-
-		dockerDevices[i] = docker.BlkioThrottleDevice{
-			Path: dev.Path,
-			Rate: rate,
-		}
-		resolvedDevices[i] = ResolvedThrottleDevice{
-			Path: dev.Path,
-			Rate: rate,
-		}
-	}
-
-	return dockerDevices, resolvedDevices
-}
-
-// hasCPUFreqSettings returns true if the resource limits have any CPU frequency settings.
-func hasCPUFreqSettings(cfg *config.ResourceLimits) bool {
-	if cfg == nil {
-		return false
-	}
-	return cfg.CPUFreq != "" || cfg.CPUTurboBoost != nil || cfg.CPUGovernor != ""
-}
-
-// buildCPUFreqConfig builds a cpufreq.Config from resource limits.
-func buildCPUFreqConfig(cfg *config.ResourceLimits) *cpufreq.Config {
-	if cfg == nil {
-		return nil
-	}
-
-	cpufreqCfg := &cpufreq.Config{
-		Frequency:  cfg.CPUFreq,
-		TurboBoost: cfg.CPUTurboBoost,
-		Governor:   cfg.CPUGovernor,
-	}
-
-	// Default governor to "performance" if frequency is set but governor isn't.
-	if cpufreqCfg.Frequency != "" && cpufreqCfg.Governor == "" {
-		cpufreqCfg.Governor = "performance"
-	}
-
-	return cpufreqCfg
-}
-
-// logCPUFreqInfo logs CPU frequency information for the target CPUs.
-func logCPUFreqInfo(log logrus.FieldLogger, mgr cpufreq.Manager, targetCPUs []int) {
-	infos, err := mgr.GetCPUInfo()
-	if err != nil {
-		log.WithError(err).Warn("Failed to get CPU frequency info")
-		return
-	}
-
-	// Filter to target CPUs if specified.
-	targetSet := make(map[int]struct{}, len(targetCPUs))
-	for _, cpuID := range targetCPUs {
-		targetSet[cpuID] = struct{}{}
-	}
-
-	for _, info := range infos {
-		// Skip CPUs not in target set if targets were specified.
-		if len(targetCPUs) > 0 {
-			if _, ok := targetSet[info.ID]; !ok {
-				continue
-			}
-		}
-
-		log.WithFields(logrus.Fields{
-			"cpu":         info.ID,
-			"min_freq":    cpufreq.FormatFrequency(info.MinFreqKHz),
-			"max_freq":    cpufreq.FormatFrequency(info.MaxFreqKHz),
-			"current":     cpufreq.FormatFrequency(info.CurrentFreqKHz),
-			"governor":    info.Governor,
-			"scaling_min": cpufreq.FormatFrequency(info.ScalingMinKHz),
-			"scaling_max": cpufreq.FormatFrequency(info.ScalingMaxKHz),
-		}).Info("CPU frequency info")
-	}
-}
-
-// runTestsWithCheckpointRestore executes tests using Podman checkpoint/restore.
-// After the initial container reaches RPC readiness, the method checkpoints the
-// container's memory state and snapshots the data directory. For each test, it
-// rolls back the data directory and restores the container from the checkpoint —
-// the client process resumes mid-execution without restart or RPC polling.
-//
-// Two data-directory rollback strategies are supported:
-//   - ZFS snapshots (when datadir.method is "zfs"): instant copy-on-write rollback.
-//   - Copy-based (when no datadir is configured): cp -a snapshot, rsync restore.
-//
-//nolint:gocognit,cyclop // Per-test checkpoint/restore is inherently complex.
-func (r *runner) runTestsWithCheckpointRestore(
-	ctx context.Context,
-	params *containerRunParams,
-	spec client.Spec,
-	containerID string,
-	containerIP string,
-	dropMemoryCaches string,
-	dropCachesPath string,
-	resultsDir string,
-	logCancel *context.CancelFunc,
-	logDone *chan struct{},
-	benchmarkoorLog *os.File,
-	cleanupFuncs *[]func(),
-	cleanupStarted chan struct{},
-) (*executor.ExecutionResult, error) {
-	log := r.log.WithFields(logrus.Fields{
-		"instance": params.Instance.ID,
-		"run_id":   params.RunID,
-		"strategy": "container-checkpoint-restore",
-	})
-
-	// Type-assert the container manager to CheckpointManager.
-	cpMgr, ok := r.containerMgr.(podman.CheckpointManager)
-	if !ok {
-		return nil, fmt.Errorf("container manager does not support checkpoint/restore")
-	}
-
-	// Resolve the test list.
-	tests := params.Tests
-	if tests == nil {
-		tests = r.executor.GetTests()
-	}
-
-	if len(tests) == 0 {
-		return &executor.ExecutionResult{}, nil
-	}
-
-	log.WithField("tests", len(tests)).Info(
-		"Running tests with checkpoint-restore strategy",
-	)
-
-	useZFS := params.DataDirCfg != nil && params.DataDirCfg.Method == "zfs"
-
-	// Find the data mount source — this is the host path the container
-	// mounts for its data directory.
-	dataMountSource := ""
-
-	containerDir := spec.DataDir()
-	if params.DataDirCfg != nil && params.DataDirCfg.ContainerDir != "" {
-		containerDir = params.DataDirCfg.ContainerDir
-	}
-
-	for _, mnt := range params.ContainerSpec.Mounts {
-		if mnt.Target == containerDir {
-			dataMountSource = mnt.Source
-
-			break
-		}
-	}
-
-	if dataMountSource == "" {
-		return nil, fmt.Errorf("could not find data mount for %s in container spec", containerDir)
-	}
-
-	// 1. Optionally restart the container before checkpointing.
-	//    Some clients (e.g., Erigon with MDBX) keep in-memory caches or
-	//    dirty state that interferes with CRIU checkpoint/restore. A
-	//    stop+start cycle gives us a cold-start process with a cleanly
-	//    shut-down database — ideal for a reliable checkpoint.
-	if r.cfg.FullConfig.GetCheckpointRestartContainer(params.Instance) {
-		log.Info("Restarting container before checkpoint for clean process state")
-
-		if err := r.containerMgr.StopContainer(ctx, containerID); err != nil {
-			return nil, fmt.Errorf("stopping container before checkpoint restart: %w", err)
-		}
-
-		// Drain logs from the stopped container.
-		waitForLogDrain(logDone, logCancel, logDrainTimeout)
-
-		if err := r.containerMgr.StartContainer(ctx, containerID); err != nil {
-			return nil, fmt.Errorf("starting container after checkpoint restart: %w", err)
-		}
-
-		// IP may change after restart; refresh it.
-		newIP, err := r.containerMgr.GetContainerIP(
-			ctx, containerID, r.cfg.DockerNetwork,
-		)
-		if err != nil {
-			return nil, fmt.Errorf("getting container IP after checkpoint restart: %w", err)
-		}
-
-		containerIP = newIP
-
-		// Restart log streaming for the restarted container.
-		if logErr := r.startLogStreaming(
-			ctx, resultsDir,
-			params.Instance.ID, containerID,
-			benchmarkoorLog, &containerLogInfo{
-				Name:             params.ContainerSpec.Name,
-				ContainerID:      containerID,
-				Image:            params.ContainerSpec.Image,
-				GenesisGroupHash: params.GenesisGroupHash,
-			},
-			params.BlockLogCollector, cleanupStarted,
-			logDone, logCancel, cleanupFuncs,
-		); logErr != nil {
-			return nil, fmt.Errorf(
-				"log streaming after checkpoint restart: %w", logErr,
-			)
-		}
-
-		// Wait for RPC readiness on the restarted container.
-		if _, err := r.waitForRPC(ctx, containerIP, spec.RPCPort()); err != nil {
-			return nil, fmt.Errorf(
-				"waiting for RPC after checkpoint restart: %w", err,
-			)
-		}
-
-		// Honour the post-RPC-ready wait if configured.
-		if waitDuration := r.cfg.FullConfig.GetWaitAfterRPCReady(
-			params.Instance,
-		); waitDuration > 0 {
-			log.WithField("duration", waitDuration).Info(
-				"Waiting after RPC ready (post-restart)",
-			)
-
-			select {
-			case <-time.After(waitDuration):
-			case <-ctx.Done():
-				return nil, ctx.Err()
-			}
-		}
-
-		log.Info("Container restarted and RPC ready, proceeding to checkpoint")
-	}
-
-	// 2. Run pre-run steps on the live container before checkpointing.
-	//    These steps (e.g., genesis setup) must be baked into the
-	//    checkpoint so every restored container starts post-pre-run.
-	engineEndpoint := fmt.Sprintf("http://%s:%d", containerIP, spec.EnginePort())
-
-	preRunOpts := &executor.ExecuteOptions{
-		EngineEndpoint: engineEndpoint,
-		JWT:            r.cfg.JWT,
-		ResultsDir:     resultsDir,
-	}
-
-	if n, err := r.executor.RunPreRunSteps(ctx, preRunOpts); err != nil {
-		return nil, fmt.Errorf("running pre-run steps before checkpoint: %w", err)
-	} else if n > 0 {
-		log.WithField("steps", n).Info("Pre-run steps completed before checkpoint")
-	}
-
-	// 3. Decide checkpoint export path: tmpfs (RAM) or disk.
-	//
-	// When checkpoint_tmpfs_threshold is configured and the container's
-	// current memory usage is under the threshold, store the checkpoint
-	// on a tmpfs mount to eliminate disk I/O during restores.
-	exportPath := filepath.Join(resultsDir, "checkpoint.tar")
-	tmpfsDir := ""
-
-	thresholdStr := r.cfg.FullConfig.GetCheckpointTmpfsThreshold(params.Instance)
-	if thresholdStr != "" {
-		threshold, parseErr := config.ParseByteSize(thresholdStr)
-		if parseErr != nil {
-			return nil, fmt.Errorf("parsing checkpoint_tmpfs_threshold: %w", parseErr)
-		}
-
-		// Read container memory usage before checkpoint.
-		reader, readerErr := stats.NewReader(
-			r.log, r.getDockerClient(), containerID,
-		)
-		if readerErr != nil {
-			log.WithError(readerErr).Warn(
-				"Failed to create stats reader for tmpfs decision, using disk",
-			)
-		} else {
-			containerStats, statsErr := reader.ReadStats()
-			_ = reader.Close()
-
-			if statsErr != nil {
-				log.WithError(statsErr).Warn(
-					"Failed to read container stats for tmpfs decision, using disk",
-				)
-			} else if containerStats.Memory > 0 && containerStats.Memory <= threshold {
-				dir, mkErr := os.MkdirTemp("", "benchmarkoor-cp-tmpfs-")
-				if mkErr != nil {
-					log.WithError(mkErr).Warn(
-						"Failed to create tmpfs dir, using disk",
-					)
-				} else {
-					// Mount tmpfs sized to the configured max, or
-					// 2x the threshold as a default.
-					tmpfsMaxSize := r.cfg.FullConfig.GetCheckpointTmpfsMaxSize(params.Instance)
-
-					var tmpfsSize uint64
-					if tmpfsMaxSize > 0 {
-						tmpfsSize = tmpfsMaxSize
-					} else {
-						tmpfsSize = threshold * 2
-					}
-
-					//nolint:gosec // Arguments are computed, not user-supplied.
-					mountCmd := exec.CommandContext(
-						ctx, "mount", "-t", "tmpfs",
-						"-o", fmt.Sprintf("size=%d", tmpfsSize),
-						"tmpfs", dir,
-					)
-					if mountOut, mountErr := mountCmd.CombinedOutput(); mountErr != nil {
-						log.WithError(mountErr).WithField(
-							"output", string(mountOut),
-						).Warn("Failed to mount tmpfs, using disk")
-
-						_ = os.Remove(dir)
-					} else {
-						tmpfsDir = dir
-						exportPath = filepath.Join(dir, "checkpoint.tar")
-
-						log.WithFields(logrus.Fields{
-							"memory_bytes":    containerStats.Memory,
-							"threshold_bytes": threshold,
-							"tmpfs_size":      tmpfsSize,
-						}).Info("Using tmpfs for checkpoint storage")
-					}
-				}
-			} else {
-				log.WithFields(logrus.Fields{
-					"memory_bytes":    containerStats.Memory,
-					"threshold_bytes": threshold,
-				}).Info("Container memory exceeds tmpfs threshold, using disk")
-			}
-		}
-	}
-
-	// 4. Checkpoint the running container.
-	//
-	// Close idle HTTP connections first so there are no ESTABLISHED TCP
-	// connections inside the container (from RPC readiness checks). CRIU
-	// refuses to checkpoint established connections without --tcp-established,
-	// and restoring them fails when the container IP changes.
-	http.DefaultClient.CloseIdleConnections()
-	time.Sleep(200 * time.Millisecond) // Let server-side sockets close.
-
-	waitAfterTCPDrop := r.cfg.FullConfig.GetCheckpointWaitAfterTCPDropConns(params.Instance)
-
-	if err := cpMgr.CheckpointContainer(ctx, containerID, exportPath, waitAfterTCPDrop); err != nil {
-		return nil, fmt.Errorf("checkpointing container: %w", err)
-	}
-
-	defer func() {
-		_ = os.Remove(exportPath)
-
-		if tmpfsDir != "" {
-			//nolint:gosec // Path is from os.MkdirTemp, not user-supplied.
-			if umountErr := exec.Command("umount", tmpfsDir).Run(); umountErr != nil {
-				log.WithError(umountErr).Warn("Failed to unmount tmpfs")
-			}
-
-			_ = os.Remove(tmpfsDir)
-		}
-	}()
-
-	// Wait for log drain after checkpoint (container has stopped).
-	waitForLogDrain(logDone, logCancel, logDrainTimeout)
-
-	// 5. Flush dirty pages and snapshot the data directory.
-	//
-	// The checkpointed process used mmap (MAP_SHARED) for its database,
-	// so dirty pages may still be in the page cache. Sync ensures they
-	// are flushed before the snapshot, so the snapshot captures the exact
-	// filesystem state that matches the checkpointed memory.
-	if syncErr := exec.Command("sync").Run(); syncErr != nil {
-		log.WithError(syncErr).Warn("Failed to sync before data directory snapshot")
-	}
-
-	// Create the data directory snapshot using either ZFS or copy-based
-	// strategy. Both are abstracted behind rollback/cleanup callbacks.
-	type snapshotRollback struct {
-		rollback func(ctx context.Context) error
-		cleanup  func()
-	}
-
-	var sr snapshotRollback
-
-	if useZFS {
-		zfsMgr := datadir.NewCheckpointZFSManager(r.log)
-
-		snapshot, snapErr := zfsMgr.SnapshotReady(ctx, &datadir.CheckpointConfig{
-			DataDir:    dataMountSource,
-			InstanceID: params.Instance.ID,
-		})
-		if snapErr != nil {
-			return nil, fmt.Errorf("creating ready-state ZFS snapshot: %w", snapErr)
-		}
-
-		sr = snapshotRollback{
-			rollback: func(ctx context.Context) error {
-				return zfsMgr.RollbackToReady(ctx, snapshot)
-			},
-			cleanup: func() {
-				if destroyErr := zfsMgr.DestroySnapshot(snapshot); destroyErr != nil {
-					log.WithError(destroyErr).Warn(
-						"Failed to destroy ready-state ZFS snapshot",
-					)
-				}
-			},
-		}
-	} else {
-		copyMgr := datadir.NewCheckpointCopyManager(r.log)
-
-		snapshot, snapErr := copyMgr.SnapshotReady(ctx, &datadir.CheckpointConfig{
-			DataDir:    dataMountSource,
-			InstanceID: params.Instance.ID,
-		})
-		if snapErr != nil {
-			return nil, fmt.Errorf("creating ready-state copy snapshot: %w", snapErr)
-		}
-
-		sr = snapshotRollback{
-			rollback: func(ctx context.Context) error {
-				return copyMgr.RollbackToReady(ctx, snapshot)
-			},
-			cleanup: func() {
-				if destroyErr := copyMgr.DestroySnapshot(snapshot); destroyErr != nil {
-					log.WithError(destroyErr).Warn(
-						"Failed to destroy ready-state copy snapshot",
-					)
-				}
-			},
-		}
-	}
-
-	defer sr.cleanup()
-
-	log.Info("Container checkpointed, starting per-test restore loop")
-
-	// 6. Per-test restore loop.
-	combined := &executor.ExecutionResult{}
-	startTime := time.Now()
-
-	for i, test := range tests {
-		select {
-		case <-ctx.Done():
-			combined.TotalDuration = time.Since(startTime)
-
-			return combined, ctx.Err()
-		default:
-		}
-
-		testLog := log.WithFields(logrus.Fields{
-			"test":  test.Name,
-			"index": fmt.Sprintf("%d/%d", i+1, len(tests)),
-		})
-
-		testLog.Info("Preparing test: rolling back data directory and restoring container")
-
-		// Flush dirty pages and drop caches BEFORE rollback. The killed
-		// container's MAP_SHARED mmap writes leave dirty pages in the
-		// page cache. Writing to drop_caches triggers a sync first, which
-		// would write those stale pages onto the rolled-back dataset —
-		// effectively undoing the rollback for recently-written blocks
-		// (e.g. MDBX meta-pages). By syncing and dropping caches before
-		// the rollback, we ensure no dirty pages survive to corrupt the
-		// post-rollback state.
-		if dropCachesPath != "" {
-			if syncErr := exec.Command("sync").Run(); syncErr != nil {
-				testLog.WithError(syncErr).Warn("Failed to sync before rollback")
-			}
-
-			if cacheErr := os.WriteFile(dropCachesPath, []byte("3"), 0); cacheErr != nil {
-				testLog.WithError(cacheErr).Warn("Failed to drop page caches before rollback")
-			}
-		}
-
-		// Roll back the data directory to the ready-state snapshot so
-		// the container restores onto clean data at the same mount path.
-		if err := sr.rollback(ctx); err != nil {
-			combined.TotalDuration = time.Since(startTime)
-
-			return combined, fmt.Errorf("rolling back data directory for test %d: %w", i, err)
-		}
-
-		// Restore container from checkpoint.
-		restoreName := fmt.Sprintf("%s-restore-%d", params.ContainerSpec.Name, i)
-
-		restoredID, err := cpMgr.RestoreContainer(ctx, exportPath, &podman.RestoreOptions{
-			Name:        restoreName,
-			NetworkName: r.cfg.DockerNetwork,
-		})
-		if err != nil {
-			combined.TotalDuration = time.Since(startTime)
-
-			return combined, fmt.Errorf("restoring container for test %d: %w", i, err)
-		}
-
-		// Register cleanup for this iteration.
-		iterID := restoredID
-
-		*cleanupFuncs = append(*cleanupFuncs, func() {
-			if rmErr := r.containerMgr.RemoveContainer(
-				context.Background(), iterID,
-			); rmErr != nil && !isContainerNotFound(rmErr) {
-				testLog.WithError(rmErr).Warn("Failed to remove restored container")
-			}
-		})
-
-		// Get container IP.
-		restoredIP, err := r.containerMgr.GetContainerIP(
-			ctx, restoredID, r.cfg.DockerNetwork,
-		)
-		if err != nil {
-			combined.TotalDuration = time.Since(startTime)
-
-			return combined, fmt.Errorf("getting container IP for test %d: %w", i, err)
-		}
-
-		// Start log streaming for restored container.
-		if logErr := r.startLogStreaming(
-			ctx, resultsDir,
-			params.Instance.ID, restoredID,
-			benchmarkoorLog, &containerLogInfo{
-				Name:             restoreName,
-				ContainerID:      restoredID,
-				Image:            params.ContainerSpec.Image,
-				GenesisGroupHash: params.GenesisGroupHash,
-			},
-			params.BlockLogCollector, cleanupStarted,
-			logDone, logCancel, cleanupFuncs,
-		); logErr != nil {
-			combined.TotalDuration = time.Since(startTime)
-
-			return combined, fmt.Errorf(
-				"log streaming for test %d: %w", i, logErr,
-			)
-		}
-
-		// No waitForRPC needed — process resumes at checkpoint state.
-		testLog.Info("Executing test (restored from checkpoint)")
-
-		// Execute single test with no executor-level rollback.
-		execOpts := &executor.ExecuteOptions{
-			EngineEndpoint: fmt.Sprintf(
-				"http://%s:%d", restoredIP, spec.EnginePort(),
-			),
-			JWT:              r.cfg.JWT,
-			ResultsDir:       resultsDir,
-			Filter:           r.cfg.TestFilter,
-			ContainerID:      restoredID,
-			DockerClient:     r.getDockerClient(),
-			DropMemoryCaches: dropMemoryCaches,
-			DropCachesPath:   dropCachesPath,
-			RollbackStrategy: config.RollbackStrategyNone,
-			RPCEndpoint: fmt.Sprintf(
-				"http://%s:%d", restoredIP, spec.RPCPort(),
-			),
-			Tests:                         []*executor.TestWithSteps{test},
-			BlockLogCollector:             params.BlockLogCollector,
-			RetryNewPayloadsSyncingConfig: r.cfg.FullConfig.GetRetryNewPayloadsSyncingState(params.Instance),
-			PostTestRPCCalls:              r.cfg.FullConfig.GetPostTestRPCCalls(params.Instance),
-		}
-
-		result, execErr := r.executor.ExecuteTests(ctx, execOpts)
-		if execErr != nil {
-			testLog.WithError(execErr).Error("Test execution failed")
-		}
-
-		// Force-remove the restored container (kills it immediately).
-		// No graceful stop needed — the test is done and the container
-		// state will be rebuilt from checkpoint for the next test.
-		if rmErr := r.containerMgr.RemoveContainer(ctx, restoredID); rmErr != nil &&
-			!isContainerNotFound(rmErr) {
-			testLog.WithError(rmErr).Warn("Failed to remove restored container")
-		}
-
-		waitForLogDrain(logDone, logCancel, logDrainTimeout)
-
-		// Aggregate results.
-		if result != nil {
-			combined.TotalTests += result.TotalTests
-			combined.Passed += result.Passed
-			combined.Failed += result.Failed
-
-			if result.StatsReaderType != "" {
-				combined.StatsReaderType = result.StatsReaderType
-			}
-
-			if result.ContainerDied {
-				combined.ContainerDied = true
-				combined.TotalDuration = time.Since(startTime)
-
-				return combined, nil
-			}
-		}
-	}
-
-	combined.TotalDuration = time.Since(startTime)
-
-	return combined, nil
-}
-
-// isContainerNotFound returns true if the error indicates the container no
-// longer exists. This is expected when cleanup runs after a container was
-// already removed (e.g., restored containers removed after each test).
-func isContainerNotFound(err error) bool {
-	return err != nil && strings.Contains(err.Error(), "no such container")
 }
